MShort

MShort.MShort — Module

Predicates that signal lowered chances of success.

MShort.find — Method

A good short should have high volume, (too) high price change, and dominating red candles.

MShort.fullret — Method

Has a full retracement occurred?

MShort.isbelow20 — Method

Is the close below the Average?

MShort.isbelow50 — Method

Is the close below the average with volume above average? Intended to be used on longer windows in contrast to the volume-less version.

MShort.isdcandles — Method

Are down candles more than up candles?

MShort.lowhin — Method

True if series experienced a recent breakout with high volume retracement (Low out, High in). If breakout is not met, returns nothing. br: Breakout rule. window: How far backward to look for breakouts. delay: Delay to look for retracements.

MShort.mustd — Method

Breakout level. Mean with std.