MVP

MVP.MVP — Module

Momentum, volume, price indicator.

MVP.discrete_mvp — Method

MVP.is_mvp — Method

Returns the mvp-ness of a pair as a sum of each condition ratio weighted by weights. If real=false it will return a Bool indicating if the pair passes the given ratios.

MVP.momentum — Method

Ratio is the minimum number of green candles required.

MVP.volume — Method

Ratio is the minimum percent increment of volume from the first half of the series.