Collections

Collections.AssetCollection — Type

A collection of assets instances, indexed by asset and exchange identifiers.

TimeTicks.DateRange — Type

Makes a daterange that spans the common min and max dates of the collection.

Collections.bqe — Method

Dispatch based on either base, quote currency, or exchange.

Collections.flatten — Method

Returns a Dict{TimeFrame, DataFrame} of all the OHLCV dataframes present in the asset collection.

Collections.iscashable — Method

Checks that all assets in the universe match the cash.