Data

Data.OHLCVTuple — Type

Similar to a StructArray (and should probably be replaced by it), used for fast conversion.

Data.ZarrInstance — Type

Candles data is stored with hierarchy PAIR -> [TIMEFRAMES...]. A pair is a ZGroup, a timeframe is a ZArray.

Base.unique! — Method

Remove duplicate from a zarray.

In a 2d zarray where we want values where the second column is unique:

unique!(x->x[2], z)

Data._load_ohlcv — Method

Load ohlcv pair data from zarr instance. za: The zarr array holding the data key: the name of the array to load from the zarr instance (in the format exchange/timeframe/pair) td: the timeframe (as integer in milliseconds) of the target ohlcv table to be loaded from, to: date range

Data._save_data — Method

Saves data to a zarr array ensuring only dates seriality, not contiguity (as opposed to _save_ohlcv).

Data.candleat — Method

Get the candle at given date from a ohlcv dataframe as a Candle.

Data.candlelast — Method

Get the last candle from a ohlcv dataframe as a Candle.

Data.candlepair — Method

Same as candleat but also fetches the previous candle, returning a Tuple{Candle, Candle}.

Data.chunksize — Method

Choose chunk size depending on size of data with a predefined split (e.g. 1/100), padding to the nearest power of 2.

Data.contiguous_ts — Method

Checks if a timeseries has any intervals not conforming to the given timeframe.

Data.df! — Method

Construct a DataFrame without copying.

Data.empty_ohlcv — Method

An empty OHLCV dataframe.

Data.key_path — Method

The full key of the data stored for the (exchange, pair, timeframe) combination.

Data.load — Method

Load a pair ohlcv data from storage. as_z: returns the ZArray

Data.load_data — Method

Load data from zarr instance. zi: The zarr instance to use key: the name of the array to load from the zarr instance (full key path). type: Set to the type that zarr should use to store the data (only bits types). [Float64]. serialized: If set, data will be deserialized before returned (type is ignored). from, to: date range sz: The chunks tuple which should match the shape of the already saved data.

Mismatching chunks

Loading data with from key with wrong dimensions (ndims(sz)) or shape (columns) will reset the stored zarray.

Data.load_ohlcv — Method

Load data from given zarr instance, exchange, pairs list and timeframe.

Data.save_ohlcv — Method

data_col: the timestamp column of the new data (1) saved_col: the timestamp column of the existing data (1) pair: the trading pair (BASE/QUOTE string) timeframe: exchange timeframe (from exc.timeframes) type: Primitive type used for storing the data (Float64) check:

:bounds (default) only checks that new data is adjacent to previous data.

:all checks full contiguity of previous and new data.

:none or anything else, no checks are done.

Data.snakecased — Method

Normalizes or special characthers separators to _.

Data.to_ohlcv — Method

Convert raw ccxt OHLCV data (matrix) to a dataframe.

Data.zdelete! — Method

Delete elements from a ZArray within the range of dates from_dt:to_dt.

Use the select function to customize the dates vector where to index th range. Defaults to the first column of a 2D Zarray.

Data.zilmdb — Function

Create a ZarrInstance at specified path using lmdb as backend.

force: resets the underlying store.

Data.@as_mat — Macro

Redefines given variable to a Matrix with type of the underlying container type.

Data.@to_mat — Macro

Same as as_mat but returns the new matrix.