Engine

Base.fill! — Method

Main.Engine.Instances.fill! all the instances with given timeframes data...

Base.fill! — Method

Pulls data from storage, or resample from the shortest timeframe available.

Data.stub! — Method

Replaces the data of the asset instances with src which should be a mapping. Used for backtesting.

src: The mapping, should be a pair TimeFrame => Dict{String, PairData}.

Example:

using Scrapers.BinanceData as bn
using Strategies
using Exchanges
setexchange!(:binanceusdm)
cfg = Config(nameof(exc.id))
strat = strategy!(:Example, cfg)
data = bn.binanceload()
stub!(strat.universe, data)