Instances

Instances.LEVERAGE_PRECISION — Constant

The number of digits allowed for leverage values.

Instances.POSITION_PRECISION — Constant

The number of digits to keep for margin calculations.

Instances.AssetInstance15 — Type

An asset instance holds all known state about an asset, i.e. BTC/USDT:

asset: the identifier

data: ohlcv series

history: the trade history of the pair

cash: how much is currently held, can be positive or negative (short)

exchange: the exchange instance that this asset instance belongs to.

limits: minimum order size (from exchange)

precision: number of decimal points (from exchange)

Instances.Position — Type

A position tracks the margin state of an asset instance:

timestamp: last update time of the position (DateTime)

side: the side of the position <:PositionSide Long or Short

tiers: sorted dict of all the leverage tiers

cash: it is the notional value of the position

For the rest of the fields refer to ccxt docs

Base.isopen — Method

Check if an asset position is open.

Base.iszero — Method

Test if some amount (base currency) is zero w.r.t. an asset instance min limit.

Base.iszero — Method

Test if asset cash is zero.

Base.iszero — Method

Test if asset cash is zero.

Base.position — Method

Asset instance long position.

Base.position — Method

Asset instance short position.

Base.position — Method

Returns the last open asset position or nothing.

Base.position — Method

Asset position by order.

Base.similar — Method

Returns a similar asset instance with cash and orders reset.

Data.candlelast — Method

Get the last available candle strictly lower than apply(tf, date)

ExchangeTypes.exchange — Method

The exchange of the asset instance.

ExchangeTypes.exchangeid — Method

ExchangeID for the asset instance.

Exchanges.leverage! — Method

Updates asset position leverage for asset instance.

Exchanges.leverage! — Method

Some exchanges consider a value of 0 leverage as max leverage for the current tier (in cross margin mode).

Exchanges.leverage! — Method

Updates position leverage.

Exchanges.tier — Method

Asset position tier.

Instances._roundpos — Function

Round function for values of position fields.

Instances.additional! — Function

Sets additional margin (should always be positive).

Instances.additional — Method

Position additional margin.

Instances.additional — Method

Asset position additional margin.

Instances.addmargin! — Method

Adds margin to a position.

Instances.asset — Method

The asset string id.

Instances.bankruptcy — Method

The price where the position is fully liquidated.

Instances.bankruptcy — Method

The price where the asset position is fully liquidated.

Instances.collateral — Method

Maximum value that can be lost by the position

Instances.committed — Method

Position locked in pending orders.

Instances.entryprice! — Function

Update the entry price.

Instances.initial! — Function

Sets initial margin (should always be positive).

Instances.instance — Function

Constructs an asset instance loading data from a zarr instance. Requires an additional external constructor defined in Engine.

Instances.isdust — Method

True if the asset value is below minimum quantity.

Instances.isdust — Method

True if the position value of the asset is below minimum quantity.

Instances.leverage — Method

Position leverage.

Instances.leverage — Method

Asset position leverage.

Instances.liqprice! — Method

Sets the liquidation price for a long position.

Instances.liqprice! — Method

Sets the liquidation price for a short position.

Instances.liqprice! — Method

Sets asset position liquidation price.

Instances.liqprice — Method

Position liquidation price.

Instances.liqprice — Method

Asset position liquidation price.

Instances.load! — Method

Load ohlcv data of asset instance.

Instances.maintenance! — Method

Sets maintenance margin.

Instances.maintenance — Method

Position maintenance margin.

Instances.maintenance — Method

Asset position maintenance margin.

Instances.makerfees — Method

Maker fees for the asset instance (usually lower than taker fees.)

Instances.margin! — Method

Sets initial margin given notional and leverage values.

Instances.margin — Method

Position initial margin (includes additional).

Instances.margin — Method

Asset position initial margin.

Instances.maxfees — Method

The maximum fees for trading in the asset market (usually the lowest vip level.)

Instances.minfees — Method

The minimum fees for trading in the asset market (usually the highest vip level.)

Instances.mmr — Method

Asset position maintenance margin rate.

Instances.mmr — Method

Position maintenance margin rate.

Instances.nondust — Function

Returns the asset cash rounded to precision.

Instances.notional! — Method

Update the notional value.

Instances.notional — Method

Position notional value.

Instances.notional — Method

Asset position notional value.

Instances.pnl — Function

Calc PNL for long position given current_price as input.

Instances.pnl — Function

Calc PNL for short position given current_price as input.

Instances.pnl — Method

The pnl of an asset position.

Instances.pnlpct — Method

Calc PNL percentage.

Instances.pnlpct — Method

The pnl percentage of an asset position.

Instances.price — Method

Asset entry price.

Instances.price — Method

Position entryprice.

Instances.price — Method

Asset entry price.

Instances.status! — Method

Opens or closes the status of an non-hedged position.

Instances.status! — Method

Opens or closes the status of an hedged position.

Instances.status — Method

Position status (open or closed).

Instances.status — Method

Asset position status (open or closed).

Instances.takerfees — Method

Taker fees for the asset instance (usually higher than maker fees.)

Instances.tier! — Function

Updates position leverage tier according to size.

Instances.timestamp — Method

Last position update time

Instruments.cash! — Method

Set position cash value.

Instruments.cash — Method

Held position.

Instruments.raw — Method

The asset string id.

Instruments.value — Method

The value held by the position, margin with pnl minus fees.

Misc.opposite — Method

The opposite position w.r.t. the asset instance and another Position or PositionSide.

Misc.reset! — Method

Resets asset cash and committments.

Misc.reset! — Method

Resets asset positions.

Misc.reset! — Method

Resets position to initial state.

Also resets leverage

When reopening a position, leverage should be set again.

Misc.reset! — Method

Resets the bare fields to close a position.

OrderTypes.commit! — Method

Set position committed cash value.

Instances.@_round — Macro

Rounds a value based on the precision field of the ai asset instance. [amount].

Instances.@ramount — Macro

Rounds a value based on the precision (amount) field of the ai asset instance.

Instances.@rprice — Macro

Rounds a value based on the precision (price) field of the ai asset instance.