Optimization

Optimization.OptSession18 — Type

An optimization session stores all the evaluated parameters combinations.

BlackBoxOptim.bboptimize — Method

Optimize parameters using the BlackBoxOptim package.

repeats: how many times to run the backtest for each step

seed: random seed

kwargs: The arguments to pass to the underlying BBO function. See the docs for the BlackBoxOptim package. Here are some most common parameters:

MaxTime: max evaluation time for the optimization

MaxFuncEvals: max number of function (backtest) evaluation

TraceMode: (:silent, :compact, :verbose) controls the logging

MaxSteps, MaxStepsWithoutProgress

From within your strategy define four ping! functions:

ping!(::Strategy, ::OptSetup): for the period of time to evaluate and the parameters space for the optimization..

ping!(::Strategy, params, ::OptRun): called before running the backtest, should apply the parameters to the strategy

Optimization._multi_opt_func — Method

Multi(threaded) optimization function.

Optimization._single_opt_func — Method

Single(threaded) optimization function.

Optimization.bbomethods — Function

Get a filtered list of methods supported by BBO (single or multi).

Optimization.define_median_func — Method

The median in multi(objective) mode has to be applied over all the (repeated) iterations.

Optimization.delete_sessions! — Method

Clear all optimization session of a strategy. keep_by: will not delete sessions that match this attributes (Dict{String, Any}). - ctx: the Context of the optimization session - params: the params (NamedTuple) of the optimization session - attrs: the config (NamedTuple) of the optimization session

Optimization.gridsearch — Method

Backtests the strategy across combination of parameters. s: The strategy. seed: random seed set before each backtest run. repeats: the amount of repetitions for each combination. save_freq: how frequently (Period) to save results, when nothing (default) saving is skipped. logging: enabled logging doshuffle: shuffle parameters combinations before iterations (random search)

Optimization.load_session — Function

Load an optimization session from storage, name:

name: strategy name

start/stop: start and stop date of the backtesting context

params_k: the first letter of every param (first.(string.(keys(sess.params))))

code: hash of params and attrs truncated to 4 chars.

Only the strategy name is required, rest is optional.

Optimization.objectives — Method

Returns the number of objectives and their type.

Optimization.optsessions — Method

Returns the zarrays storing all the optimization session over the specified zarrinstance.

Optimization.progsearch — Method

A progressive search performs multiple grid searches with only 1 repetition per parameters combination. After each search is completed, the results are filtered according to custom rules. The parameters from the results that match the filtering will be backtested again with a different offset which modifies the backtesting period. rounds: how many iterations (of grid searches) to perform sess: If a Ref{<:OptSession} is provided, search will resume from the session previous results halve: At each iteration

Additional kwargs are forwarded to the grid search.

Optimization.remove_incomplete! — Method

Remove results that don't have all the repeated evalutaion.

Optimization.result_params — Function

Fetch the named tuple of a single parameters combination.

Optimization.save_session — Method

Save the optimization session over the provided zarr instance.

sess: the OptSession from: save the optimization results starting from the specified index (when saving progressively) to: save the optimization results up to the specified index (when saving progressively)

Optimization.slidesearch — Method

Backtests by sliding over the backtesting period, by the smallest timeframe (the strategy timeframe). Until a full range of timeframes is reached between the strategy timeframe and backtesting context timeframe. multiplier: the steps count (total stepps will be multiplier * context_timeframe / s.timeframe )

Strategies.ping! — Method

This ping function should apply the parameters to the strategy, called before the backtest is performed.

Strategies.ping! — Method

Has to return a Optimizations.ContextSpace named tuple where ctx (Executors.Context) is the time period to backtest and the space is either an already constructed subtype of BlackBoxOptim.SearchSpace or a tuple (Symbol, args...) for a search space pre-defined within the BBO package.