Processing

Processing.cleanup_ohlcv_data — Method

Similar to the freqtrade homonymous function.

fill_missing: :close fills non present candles with previous close and 0 volume, else with NaN.

Processing.fill_missing_candles — Method

Assuming timestamps are sorted, returns a new dataframe with a contiguous rows based on timeframe. Rows are filled either by previous close, or NaN.

Processing.iscomplete — Method

Checks if a candle is old enough to be complete.

Processing.isincomplete — Method

Checks if a candle is too new.

Processing.islast — Method

Checks if a candle is exactly the latest candle.

Processing.isleftadj — Method

a is left adjacent to b if in order ..ab..

Processing.isrightadj — Method

a is right adjacent to b if in order ..ba..

Processing.maptf — Method

Apply a function over data, resampling data to each timeframe in tfs.

f: signature is (data; kwargs...)::DataFrame

tfsum: sum the scores across multiple timeframes for every pair.

Processing.normalize! — Method

Applies either a unitrange transform or a zscore tranform over the data in place.

Processing.resample — Function

Resamples ohlcv data from a smaller to a higher timeframe.

style: how to modify the data, (arguments to the grouped dataframe) [:ohlcv]

Processing.resample — Method

Resamples data, and saves to storage.

Usually not worth it

Resampling is quite fast, so it is simpler to keep only the smaller timeframe on storage, and resample the longer ones on demand.

Processing.trail! — Method

Appends empty candles to df up to to datetime (excluded). cap: max capacity of df