Python

Provides python bindings, some helper functions and async management.

Python.Python — Module

Defines the Python module which sets up the Python interpreter and imports required modules and constants.

Python.CALLBACKS — Constant

An array of callback functions.

Python.PYMODPATHS — Constant

An array of Python module paths.

Python.PYTHONPATH — Constant

The Python module search path.

Python.PY_V — Constant

The Python version.

Python._INITIALIZED — Constant

This constant indicates whether the Python runtime has been initialized. It is used to delay initialization until the first Python call.

Python.pyCached — Constant

A dictionary for caching Python objects.

Python.pyisvalue_func — Constant

A Python object used for checking if a value is a Python object.

Python.pytryfloat — Constant

A Python object used for converting values to float.

Python.PythonAsync — Type

A structure for handling Python's asynchronous operations.

pyaio

pythreads

pyloop

pycoro_type

globs

start_func

task

task_running

This structure is used to manage Python's asynchronous operations. It contains fields for Python's asyncio, threads, event loop, coroutine type, global variables, start function, and task status.

PythonCall.PyDict — Method

PyDict(p::Pair)

Create a Python dictionary from a pair.

Base.copyto! — Method

Copies a python async structures.

copyto!(
    pa_to::Python.PythonAsync,
    pa_from::Python.PythonAsync
) -> Bool

Python._async_init — Method

Initialized a PythonAsync structure (which holds a reference to the event loop.)

_async_init(pa::Python.PythonAsync)

Python._isfutdone — Method

Checks if a Python future is done.

_isfutdone(fut::Py) -> Bool

Python._pyfetch — Method

Fetches the result of a Julia function call synchronously.

_pyfetch(
    f::Function,
    args...;
    coro_running,
    kwargs...
) -> Any

Python._pyfetch — Method

Fetches the result of a Python function call synchronously and returns an exception if any.

_pyfetch(
    f::Py,
    ::Val{:try},
    args...;
    coro_running,
    kwargs...
) -> Any

Python._pyfetch — Method

Fetches the result of a Python function call synchronously.

_pyfetch(f::Py, args...; coro_running, kwargs...) -> Any

Python._pyfetch_timeout — Method

Fetches the result of a Python function call synchronously with a timeout. If the timeout is reached, it calls another function and returns its result.

_pyfetch_timeout(
    f1::Py,
    f2::Union{Function, Py},
    timeout::Dates.Period,
    args...;
    coro_running,
    kwargs...
) -> Any

Python.async_start_runner_func! — Method

Main async loop function, sleeps indefinitely and closes loop on exception.

async_start_runner_func!(pa) -> Py

Python.clearpypath! — Method

Remove wrong python version libraries dirs from python loading path.

Python.isinitialized_async — Method

Checks if python async state (event loop) is initialized.

isinitialized_async(pa::Python.PythonAsync) -> Bool

Python.py_except_name — Method

py_except_name(e::PyException)

Get the name of a Python exception.

Python.py_start_loop — Function

Starts a python event loop, updating pa.

py_start_loop()
py_start_loop(pa::Python.PythonAsync)

Python.pycancel — Method

Cancels a Python future.

pycancel(fut::Py) -> Bool

Python.pydicthash — Method

pydicthash(d)

Calculate the hash of a Python dictionary.

Python.pyisnonzero — Method

pyisnonzero(v::Py)::Bool

Check if a Python value is nonzero.

Python.pylist_to_matrix — Method

pylist_to_matrix(data::Py)

Convert a Python list to a Julia matrix.

Python.pyloop_stop_fn — Method

Generates a function that terminates the python even loop.

pyloop_stop_fn() -> Python.var"#fn#9"

Python.pyschedule — Method

Schedules a Python coroutine to run on the event loop.

pyschedule(coro::Py) -> Union{Nothing, Py}

Python.pytask — Method

Creates a Julia task from a Python coroutine and runs it asynchronously.

pytask(coro::Py, ::Val{:coro}; coro_running) -> Task

Python.pytask — Method

Creates a Julia task from a Python coroutine and returns the Python future and the Julia task.

pytask(
    coro::Py,
    ::Val{:fut};
    coro_running
) -> Tuple{Py, Task}

Python.pytask — Method

Creates a Julia task from a Python function call and returns the Python future and the Julia task.

pytask(
    f::Py,
    ::Val{:sched},
    args...;
    coro_running,
    kwargs...
) -> Tuple{Py, Task}

Python.pytask — Method

Creates a Julia task from a Python function call and runs it asynchronously.

pytask(
    f::Py,
    args...;
    coro_running,
    kwargs...
) -> Tuple{Py, Task}

Python.pytofloat — Method

pytofloat(v::Py, def::T)::T where {T<:Number}

Convert a Python value to a Julia float, with a default value.

Python.pywait_fut — Method

Waits for a Python future to be done.

pywait_fut(fut::Py)

Python.@pymodule — Macro

Import a python module over a variable defined in global scope.

Python.@pystr — Macro

@pystr(k, v=nothing)

Convert a Julia value to a Python string representation.