Stats

Stats.resample_trades — Function

Aggregates all trades of a strategy in a single dataframe

byinstance: (trades_df, ai) -> nothing can modify the dataframe of a single instance before it is appended to the full df.

Stats.resample_trades — Method

Resamples trades data from a smaller to a higher timeframe.

Stats.trades_balance — Function

Plots the trade history for all the assets in a strategy.

return_all[true]: similar to the function for single assets, plus:

cum_quote: the balance of cash for each timestamp

cum_value_balance: the balance of all held assets in quote currency for each timestamp.

byasset[false]: also return a column that tracks the value balance by asset for each timestamp normalize_timeframes:

Stats.trades_balance — Method

Replays the trade history of a single asset instance.

return_all: if true returns a dataframe where:

base/quote_balance the volume generated by the trades that happened at that timestamp.

:cum_total represents the total balance held for each timestamp.

:cum_value_balance represents the value in quote currency of the asset for each timestamp.

For single assets only

If your strategy trades multiple assets the profits returned by this function won't match the strategy actual holdings since calculation are done only w.r.t this single asset.

Stats.tradesvolume! — Method

Buys substract quote currency, while sells subtract base currency