Strategies

Strategies.Strategy — Type

The strategy is the core type of the framework.

The strategy type is concrete according to:

Name (Symbol)

Exchange (ExchangeID), read from config

Quote cash (Symbol), read from config

The exchange and the quote cash should be specified from the config, or the strategy module.

universe: All the assets that the strategy knows about

holdings: assets with non zero balance.

buyorders: active buy orders

sellorders: active sell orders

timeframe: the smallest timeframe the strategy uses

cash: the quote currency used for trades

cash_committed: cash kept busy by pending orders

config: The Config the strategy was instantiated with.

logs: logs exchange events like positions updates.

Conventions for strategy defined attributes:

NAME: the name of the strategy could be different from module name

S: the strategy type.

TF: the smallest timeframe that the strategy uses

Instruments.freecash — Method

Cash that is not committed, and therefore free to use for new orders.

Misc.execmode — Method

Returns the strategy execution mode.

Misc.marginmode — Method

Get the strategy margin mode.

Misc.reset! — Function

Resets strategy state. defaults: if true reapply strategy config defaults.

Strategies.assets — Method

Assets loaded by the strategy.

Strategies.default! — Method

Set strategy defaults.

Strategies.exchange — Method

Strategy main exchange id.

Strategies.instances — Method

Strategy assets instance.

Strategies.ping! — Method

Called on each timestep iteration, possible multiple times. Receives:

current_time: the current timestamp to evaluate (the current candle would be current_time - timeframe).

ctx: The context of the executor.

ping!(
    ::Strategies.Strategy,
    current_time,
    ctx,
    args...;
    kwargs...
) -> OrderTypes.OrderError

Strategies.ping! — Method

When an order is cancelled the strategy is pinged with an order error.

ping!(
    ::Strategies.Strategy,
    ::OrderTypes.Order,
    err::OrderTypes.OrderError,
    ::Instances.AssetInstance15;
    kwargs...
) -> OrderTypes.OrderError

Strategies.ping! — Method

Called at the end of the reset! function applied to a strategy.

ping!(_::Strategies.Strategy, _::Strategies.ResetStrategy)

Strategies.ping! — Method

How much lookback data the strategy needs.

ping!(
    s::Strategies.Strategy,
    _::Strategies.WarmupPeriod
) -> Any

Strategies.ping! — Method

Called to construct the strategy, should return the strategy instance.

ping!(
    _::Type{<:Strategies.Strategy},
    cfg,
    _::Strategies.LoadStrategy
)

Strategies.reload! — Method

Reloads ohlcv data for assets already present in the strategy universe.

Strategies.tradesedge — Method

Returns the first and last trade of any asset in the strategy universe.