Misc

Misc.config — Constant

Global configuration instance.

Misc.fiatnames — Constant

A list of fiat and fiat-like assets names.

Misc.futures_exchange — Constant

Some exchanges are split into different classes in ccxt.

Misc.ohlcv_limits — Constant

Exchange ohlcv candles limits.

Misc.results — Constant

Holds recently evaluated statements.

Misc.tf_win — Constant

Mapping of timeframes to default window sizes.

Misc.Config — Type

The config main structure:

window: The default number of candles (OHLCV).

timeframe: The default timeframe of the candles.

qc: The default quote currency.

margin: If margin is enabled, only margin pairs are considered.

leverage:

:yes : leveraged pairs will not be filtered.

:only : ONLY leveraged will not be filtered.

:from : Selects non leveraged pairs, that also have a leveraged siblings.

futures: Selects the futures version of an Exchange.

slope/min/max: Used in Analysios/slope.

ct: Used in Analysis/corr.

attrs: Generic metadata container.

Misc._empty_df — Method

An empty OHLCV dataframe.

Misc._from_to_dt — Method

Converts integers to relative datetimes according to timeframe duration.

Misc._instantiate_workers — Method

Instantiate new workers if the current number mismatches the requested one.

Misc.clearpypath! — Method

Remove wrong python version libraries dirs from python loading path.

Misc.config_path — Method

The config path (TOML), relative to the current project directory.

Misc.loadconfig — Method

Parses the toml file and populates the global config.

Misc.printn — Function

Print a (currency) number.

Misc.resetconfig! — Method

Reset global config to default values.

Misc.setcfg! — Method

Sets a single config value.

Misc.@ifundef — Macro

Define a new symbol with given value if it is not already defined.

Misc.@lev! — Macro

Toggle config leverage flag

Misc.@margin! — Macro

Toggle config margin flag.

Misc.@pymodule — Macro

Import a python module over a variable defined in global scope.

Progress bar

Pbar.Pbar — Module

Python enlighten progressbar wrapper.

Pbar.pbclose — Method

Terminates the progress bar.

Pbar.@pbar! — Macro

Instantiate a progress bar:

data: length(data) determines the bar total

unit: what unit the display

desc: description will appear over the progressbar

Pbar.@pbupdate! — Macro

Single update to the progressbar with the new value.