Data

The main backend currently is Zarr. Zarr is similar to feather or parquet in that it optimizes to columnar data, or in general arrays. However it is simpler, and allows to pick different encoding schemes, and supports compression by default. More over the zarr interface can be backed by different storage layers, that can also be over the network. Compared to no/sql databases columnar storage has the drawback of having to read chunks for queries, but we are almost never are interested in scalar values, we always query a time-series of some sort, so the latency loss is a non issue.

Data.OHLCVTuple — Type

Similar to a StructArray (and should probably be replaced by it), used for fast conversion.

Data.ZarrInstance — Type

Candles data is stored with hierarchy PAIR -> [TIMEFRAMES...]. A pair is a ZGroup, a timeframe is a ZArray.

Base.unique! — Method

Remove duplicate from a zarray.

In a 2d zarray where we want values where the second column is unique:

unique!(x->x[2], z)

Data._load_ohlcv — Method

Load ohlcv pair data from zarr instance. zi: The zarr instance to use key: the name of the array to load from the zarr instance (in the format exchange/timeframe/pair) td: the timeframe (as integer in milliseconds) of the target ohlcv table to be loaded from, to: date range

Data._save_data — Method

Saves data to a zarr array ensuring only dates seriality, not contiguity (as opposed to _save_ohlcv).

Data.combine_data — Method

combinerows of two (OHLCV) dataframes over using :timestamp column as index.

Data.combinerows — Method

(Right)Merge two dataframes on key, assuming the key is ordered and unique in both dataframes.

Data.contiguous_ts — Method

Checks if a timeseries has any intervals not conforming to the given timeframe.

Data.data_td — Method

The time interval of the dataframe, guesses from the difference between the first two rows.

Data.empty_ohlcv — Method

An empty OHLCV dataframe.

Data.load — Method

Load a pair ohlcv data from storage. as_z: returns the ZArray

Data.load_data — Method

Load data from zarr instance. zi: The zarr instance to use key: the name of the array to load from the zarr instance (full key path). type: Set to the type that zarr should use to store the data (only bits types). [Float64]. serialized: If set, data will be deserialized before returned (type is ignored). from, to: date range

Data.load_ohlcv — Method

Load data from given zarr instance, exchange, pairs list and timeframe.

Data.pair_key — Method

The full key of the data stored for the (exchange, pair, timeframe) combination.

Data.sanitize_pair — Method

Normalizes or special characthers separators to _.

Data.save_ohlcv — Method

data_col: the timestamp column of the new data (1) saved_col: the timestamp column of the existing data (1) pair: the trading pair (BASE/QUOTE string) timeframe: exchange timeframe (from exc.timeframes) type: Primitive type used for storing the data (Float64)

Data.to_ohlcv — Method

Convert raw ccxt OHLCV data (matrix) to a dataframe.

Data.zdelete! — Method

Delete elements from a ZArray within the range of dates from_dt:to_dt.

Use the select function to customize the dates vector where to index th range. Defaults to the first column of a 2D Zarray.

Data.zilmdb — Function

Create a ZarrInstance at specified path using lmdb as backend.

force: resets the underlying store.

Data.@as_mat — Macro

Redefines given variable to a Matrix with type of the underlying container type.

Data.@as_ts — Macro

Converts a subset of columns to timeseries.

Data.@ohlc — Macro

Converts ohlcv dataframe to timeseries type.

Data.@otime — Macro

Get the last date of loaded data (in mrkts variable).

Data.@to_mat — Macro

Same as as_mat but returns the new matrix.

Prices

Main.Prices.gprofit — Function

Total profit of a ladder of trades

Main.Prices.price_ranges — Function

Given a price, output price at given ratios. (Predefined ratios - :bal,:futures,:tight,:expa)

Main.Prices.price_ranges — Method

Get the price range of a map of pairs, using the last available close price.