Analysis

Analysis.fltsummary — Method

Return the summary of a filtered vector of pairdata.

Analysis.mark! — Method

Loads the Mark module.

Base.filter — Method

Filters a list of pairs using a predicate function. The predicate functions must return a Real number which will be used for sorting.