Data

Data.ZarrInstance — Type

Candles data is stored with hierarchy PAIR -> [TIMEFRAMES...]. A pair is a ZGroup, a timeframe is a ZArray.

Data.combine_data — Method

combinerows of two (OHLCV) dataframes over using :timestamp column as index.

Data.combinerows — Method

(Right)Merge two dataframes on key, assuming the key is ordered and unique in both dataframes.

Data.contiguous_ts — Method

Checks if a timeseries has any intervals not conforming to the given timeframe.

Data.data_td — Method

The time interval of the dataframe, guesses from the difference between the first two rows.

Data.load_pair — Function

Load a pair ohlcv data from storage. as_z: returns the ZArray

Data.load_pairs — Method

Load data from given zarr instance, exchange, pairs list and timeframe.

Data.pair_key — Method

The full key of the data stored for the (exchange, pair, timeframe) combination.

Data.save_pair — Method

data_col: the timestamp column of the new data (1) saved_col: the timestamp column of the existing data (1) kind: what type of trading data it is, (ohlcv or trades) pair: the trading pair (BASE/QUOTE string) timeframe: exchange timeframe (from exc.timeframes) type: Primitive type used for storing the data (Float64)

Data.to_df — Method

Convert ccxt OHLCV data to a timearray/dataframe.

Data.@as_mat — Macro

Redefines given variable to a Matrix with type of the underlying container type.

Data.@to_mat — Macro

Same as as_mat but returns the new matrix.

Prices

Main.Prices.gprofit — Function

Total profit of a ladder of trades

Main.Prices.price_ranges — Function

Given a price, output price at given ratios. (Predefined ratios - :bal,:futures,:tight,:expa)

Main.Prices.price_ranges — Method

Get the price range of a map of pairs, using the last available close price.