Exchanges

Exchanges.aprice — Method

Get price from ticker.

Exchanges.as_spot_ticker — Method

Trims the settlement currency in futures.

Exchanges.deleverage_pair — Method

Remove leveraged pair pre/suffixes from base currency.

Exchanges.get_markets — Method

Get the the markets of the ccxt instance, according to min_volume and quote currency.

Exchanges.get_pairlist — Method

Get the exchange pairlist. quot: Only choose pairs where the quot currency equals quot. min_vol: The minimum volume of each pair. skip_fiat: Ignore fiat/fiat pairs. margin: Only choose pairs enabled for margin trading. leveraged: If :no skip all pairs where the base currency matches the leverage_pair_rgx regex. as_vec: Returns the pairlist as a Vector instead of as a Dict.

Exchanges.getexchange — Method

Get ccxt exchange by symbol.

Exchanges.hastickers — Method

Check if exchange has tickers list.

Exchanges.is_fiat_pair — Method

Check if both base and quote are fiat currencies.

Exchanges.is_leveraged_pair — Method

Test if pair has leveraged naming.

Exchanges.loadmarkets! — Method

Load exchange markets:

cache: rely on storage cache

agemax: max cache valid period [1 day].

Exchanges.price_ranges — Method

Get price ranges using tickers data from exchange.

Exchanges.pyexchange — Function

Exchanges.sanitize_pair — Method

Normalizes or special characthers separators to _.

Exchanges.setexchange! — Method

Instantiate an exchange struct. it sets:

The matching ccxt class.

Pre-emptively loads the markets.

Sets the exchange timeframes.

Sets exchange api keys.

Exchanges.@exchange! — Macro

Define an exchange variable set to its matching exchange instance.

Exchanges.@tickers — Macro

Fetch and cache tickers data.

Data

Extensions for the top level Data module.

Data.load_pairs — Method

Load all pairs from exchange according to config quote currency and timeframe.

Data.save_pair — Method

Updates pair data of the globally set exchange.

Fetching

Helper module for downloading data off exchanges.

Main.Fetch.fetch_pairs — Method

Fetch ohlcv data for multiple exchanges on the same timeframe. It accepts: - a mapping of exchange instances to pairlists. - a vector of symbols for which an exchange instance will be instantiated for each element, and pairlist will be composed according to quote currency and min_volume from Backtest.config.