Engine

Within the PingPong "model" we use ping! and pong! functions to communicate between strategies and executors. The executor "pings" the strategy, impliying that the strategy should do, or return something. The strategy instead "pongs" the executor, expecting it to do, or return something.

The user of the bot, generally, only writes ping! functions within their strategies. In the case the user requires custom behaviour that is not implemented by the framework, they might be required to write pong! functions.

Other trading bots offer a set of methods that the user can implement for tuning purposes, usually tied to the super class of the strategy. Within PingPong instead, our convention is to only deal with ping! and pong! functions, such that you know that whenever a pong! call is done from the strategy, that is a point of possible divergence between simulation and live execution.

In fact, ping and pong functions are implemented such that they dispatch differently according to the execution mode of the strategy.

There are 3 execution modes:

Sim: what the backtester uses to run the simulations

Paper: the dry run mode, that runs the bot like it would in live, working with live data feeds and simulating order execution with live prices.

Live: like Paper but with order execution being actually forwarded to CCXT.

Therefore if the strategy is instantiated in Sim mode, calling pong!(s, ...), where s is the strategy object of type Strategy{Sim, S, E}, the pong! function will dispatch to the Sim execution method. S and E are the other two parameters which a strategy type requires for concretization.

S<:Symbol: the symbol that matches the module name of the strategy, like :Example

E<:ExchangeID: The symbol already checked to match a valid CCXT exchange, which will be the exchange that the strategy will operate on.

To realize the pong! dispatch convention, you can expect the first argument of every pong function to be the strategy object itself, while ping function might have either the strategy object or the type of the strategy as first argument (Type{Strategy{...}}).

API

Base.fill! — Method

Main.Engine.Instances.fill! all the instances with given timeframes data...

Base.fill! — Method

Pulls data from storage, or resample from the shortest timeframe available.

Data.stub! — Method

Replaces the data of the asset instances with src which should be a mapping. Used for backtesting.

src: The mapping, should be a pair TimeFrame => Dict{String, PairData}.

Example:

using Scrapers.BinanceData as bn
using Strategies
using Exchanges
setexchange!(:binanceusdm)
cfg = Config(nameof(exc.id))
strat = strategy!(:Example, cfg)
data = bn.binanceload()
stub!(strat.universe, data)

Instances.AssetInstance50 — Type

An asset instance holds all known state about an asset, i.e. BTC/USDT:

asset: the identifier

data: ohlcv series

history: the trade history of the pair

cash: how much is currently held, can be positive or negative (short)

exchange: the exchange instance that this asset instance belongs to.

limits: minimum order size (from exchange)

precision: number of decimal points (from exchange)

Base.similar — Method

Returns a similar asset instance with cash and orders reset.

Data.candlelast — Method

Get the last available candle strictly lower than apply(tf, date)

Instances.load! — Method

Load ohlcv data of asset instance.

Collections.AssetCollection — Type

A collection of assets instances, indexed by asset and exchange identifiers.

TimeTicks.DateRange — Type

Makes a daterange that spans the common min and max dates of the collection.

Collections.bqe — Method

Dispatch based on either base, quote currency, or exchange.

Collections.flatten — Method

Returns a Dict{TimeFrame, DataFrame} of all the OHLCV dataframes present in the asset collection.

Collections.iscashable — Method

Checks that all assets in the universe match the cash.