Exchanges

Exchange Types

Basic exchange types, and global exchange vars.

API

Exchange functions

Data

Extensions for the top level Data module.

Data.load_ohlcv — Method

Loads all pairs given timeframe matching global exc and config.

Data.load_ohlcv — Method

Loads all pairs for given exc/timeframe matching global config and zi (ZarrInstance).

Data.load_ohlcv — Method

Load given pairs from global exc and ZarrInstance.

Data.load_ohlcv — Method

Load all pairs from exchange according to config quote currency and timeframe.

Data.save_ohlcv — Method

Updates pair data of the globally set exchange.

Fetching

Helper module for downloading data off exchanges.

Fetch.SINCE_MIN_PERIOD — Constant

Used to slide the since param forward when retrying fetching (in case the requested timestamp is too old).

Base.convert — Method

This is the fastest (afaik) way to convert ccxt lists to dataframe friendly format.

Fetch._fetch_loop — Method

Calls the fetc_func iteratively until the full dates range has been downloaded. NOTE: The returned data won't be exactly the number of candles expected by e.g. length(DateRange(from, to))

Fetch._fetch_ohlcv_from_to — Method

Ensure a to date is set, before fetching.

Fetch._fetch_with_delay — Method

Wraps a fetching function around error handling and backoff delay. fetch_func signature: (pair::String, since::Option{Float}, limit::Float) -> PyList The converter function has to tabulate the data such that the first column is the timestamp

Fetch.fetch_ohlcv — Method

Fetch ohlcv data from exchange for a list of pairs.

from, to: Can represent a date. A negative from number implies fetching the last N=from candles.

update: If true, will check for cached data, and fetch only missing candles. (false)

progress: if true, show a progress bar. (true)

reset: if true, will remove cached data before fetching. (false)

Fetch.fetch_ohlcv — Method

Fetch ohlcv data for multiple exchanges on the same timeframe.

It accepts:

a mapping of exchange instances to pairlists.

a vector of symbols for which an exchange instance will be instantiated for each element, and pairlist will be composed according to quote currency and min_volume from PingPong.config.

Fetch.find_since — Method

Should return the oldest possible timestamp for a pair, or something close to it.

Fetch.funding_data — Method

Retrieves all funding data return by exchange for symbol, or a subset.

funding_data(exc, "BTC/USDT:USDT")
funding_data(exc, "BTC/USDT:USDT", :fundingRate, :markPrice)

Fetch.funding_history — Method

Fetch funding rate history from exchange for a list of Derivative pairs.

from, to: specify date period to fetch candles for.