Configuration

API

Misc.config — Constant

Global configuration instance.

Misc.fetch_limits — Constant

Exchange ohlcv candles limits.

Misc.futures_exchange — Constant

Some exchanges are split into different classes in ccxt.

Misc.results — Constant

Holds recently evaluated statements.

Misc.Config — Type

The config main structure:

path: File path that loaded this config.

mode: Execution mode (Sim, Paper, Live)

exchange: A symbol to instantiate an exchange (a raw ExchangeID symbol)

qc: The default quote currency.

margin: If margin is enabled, only margin pairs are considered.

leverage:

:yes : Leveraged pairs will not be filtered.

:only : ONLY leveraged will not be filtered.

:from : Selects non leveraged pairs, that also have a leveraged siblings.

futures: Selects the futures version of an Exchange and/or markets.

min_vol: A minimum acceptable volume, e.g. for filtering markets.

initial_cash: Starting cash, used when instantiating a strategy.

min_size: Default order size.

min_timeframe: The default (shortest) timeframe of the candles.

timeframes: Vector of sorted timeframes that the strategy uses (for loading data).

window: (deprecated) The default number of candles (OHLCV).

attrs: Generic metadata container.

sources: mapping of modules symbols name to (.jl) file paths

Base.copy — Method

Shallow copies the config, and top level containers fields timeframes and attrs.

Base.empty! — Method

Reset config to default values.

Misc._instantiate_workers — Method

Instantiate new workers if the current number mismatches the requested one.

Misc.after — Method

Returns a view of the sorted vector v, indexed using rangeafter.

Misc.before — Method

Returns a view of the sorted vector v, indexed using rangebefore.

Misc.between — Method

Returns a view of the sorted vector v, indexed using rangebetween.

julia> between([1, 2, 3, 3, 3], 3, 3; strict=true)
0-element view(::Vector{Int64}, 6:5) with eltype Int64
julia> between([1, 2, 3, 3, 3], 1, 3; strict=true)
1-element view(::Vector{Int64}, 2:2) with eltype Int64:
 2
julia> between([1, 2, 3, 3, 3], 2, 3; strict=false)
2-element view(::Vector{Int64}, 3:4) with eltype Int64:
 3
 3

Misc.config! — Method

Parses the toml file and populates the config cfg (defaults to global config).

Misc.config_path — Method

The config path (TOML), relative to the current project directory.

Misc.rangeafter — Method

Returns the range index of sorted vector v for all the values after d. when strict is false, the range will start after the first occurence of d.

Misc.rangebefore — Method

Returns the range index of sorted vector v for all the values before d. when strict is false, the range will start before the last occurence of d.

Misc.rangebetween — Method

Returns the range index of sorted vector v for all the values before d. Argument strict behaves same as rangeafter and rangebefore.

Misc.@as_dfdict — Macro

Binds a mrkts variable to a Dict{String, DataFrame} where the keys are the pairs names and the data is the OHLCV data of the pair.

Misc.@lev! — Macro

Toggle config leverage flag

Misc.@margin! — Macro

Toggle config margin flag.

Progress bar

Pbar.Pbar — Module

Progress bar wrapper.

Pbar.frequency! — Method

Set the update frequency globally.

Pbar.pbclose! — Function

Terminates the progress bar.

Pbar.transient! — Function

Toggles pbar transient flag

Pbar.@pbar! — Macro

Instantiate a progress bar:

data: length(data) determines the bar total

unit: what unit the display

desc: description will appear over the progressbar

Pbar.@pbupdate! — Macro

Single update to the progressbar with the new value.

Pbar.@withpbar! — Macro

Same as @pbar! but with implicit closing.