Strategy interface

Load a strategy

The strategy is instantiated by loading a julia module at runtime.

using PingPong
cfg = Config(:kucoin) # Load the configuration, choosing kucoin as exchange
strategy!(:Example, cfg) # Load the Example strategy

The strategy is looked up inside the config under the sources key:

# Example config
[kucoin]
futures = true
[sources]
Example = "cfg/strategies/Example.jl" # the name of the module

The key is the name of the module (in this case Example) which will be imported from the included file "cfg/strategies/Example.jl".

After the strategy module is imported the strategy is instantiated by calling the ping!(::Type{S}, ::LoadStrategy, cfg) function.

> typeof(s)
Engine.Strategies.Strategy37{:Example, ExchangeTypes.ExchangeID{:kucoin}(), :USDT}

See here how the load method is defined.

module Example
using PingPong
@strategyenv!

const NAME = :Example
const EXCID = ExchangeID(:phemex)
const S{M} = Strategy{M,NAME,typeof(EXCID),NoMargin,:USDT}
const TF = tf"1m"

function ping!(::Type{S}, ::LoadStrategy, config)
    assets = marketsid(S)
    s = Strategy(Example, assets; config)
    s
end

end

See that the load method dispatches on the strategy type with cfg as argument of type Misc.Config.

As a rule of thumb if the method should be called before the strategy is construct, then it dispatches to the strategy type (Type{S}), otherwise the strategy instance (S). For convention the module property S of your strategy module, declares the strategy type (const S = Strategy{name, exc, ...}).

Strategy interface

Both ping! and pong! functions adhere to a convention for function signatures. The first argument is always either an instance of the subject or its type, followed by the arguments of the function, with the last non kw argument being the verb which describes the purpose of the function. KW arguments are optional and don't have any requirements. We can see below that Type{S} is the subject, config is an argument, and ::LoadStrategy is the _verb.

List of strategy ping! functions

Strategies.ping! — Function

Called on each timestep iteration, possible multiple times. Receives:

current_time: the current timestamp to evaluate (the current candle would be current_time - timeframe).

ctx: The context of the executor.

ping!(
    _::Strategies.Strategy,
    current_time::Dates.DateTime,
    ctx
)

Called to construct the strategy, should return the strategy instance.

ping!(
    _::Type{<:Strategies.Strategy},
    cfg,
    _::Strategies.LoadStrategy
)

Called at the end of the reset! function applied to a strategy.

ping!(_::Strategies.Strategy, _::Strategies.ResetStrategy)

How much lookback data the strategy needs.

ping!(
    s::Strategies.Strategy,
    _::Strategies.WarmupPeriod
) -> Any

When an order is cancelled the strategy is pinged with an order error.

ping!(
    ::Strategies.Strategy,
    ::OrderTypes.Order,
    err::OrderTypes.OrderError,
    ::Instances.AssetInstance15;
    kwargs...
) -> OrderTypes.OrderError

After a position was updated from a trade.

ping!(
    _::Strategies.Strategy{X, N, <:ExchangeID, <:Misc.WithMargin, C} where {X<:Misc.ExecMode, N, C},
    ai,
    trade::OrderTypes.Trade,
    _::Instances.Position,
    _::Instances.PositionChange
)

After a position update from a candle.

ping!(
    _::Strategies.Strategy{X, N, <:ExchangeID, <:Misc.WithMargin, C} where {X<:Misc.ExecMode, N, C},
    ai,
    date::Dates.DateTime,
    _::Instances.Position,
    _::Instances.PositionUpdate
)

Has to return a Optimizations.ContextSpace named tuple where ctx (Executors.Context) is the time period to backtest and the space is either an already constructed subtype of BlackBoxOptim.SearchSpace or a tuple (Symbol, args...) for a search space pre-defined within the BBO package.

This ping function should apply the parameters to the strategy, called before the backtest is performed.