Types

By learning the main types you get to know the building blocks to start composing your strategy for backtesting and/or live trading.

API

Instruments.Asset5 — Type

An Asset represents a parsed raw (usually ccxt) pair of base and quote currency.

raw: The raw underlying string e.g. 'BTC/USDT'

bc: base currency (Symbol)

qc: quote currency (Symbol)

fiat: if both the base and quote currencies match a known fiat symbol e.g. 'USDT/USDC'

leveraged: if parsing matched a leveraged token e.g. 'ETH3L/USDT' or 'ETH3S/USDT'

unleveraged_bc: a leveraged token with the mod removed, e.g. ETH3L => ETH

> asset = a"BTC/USDT"
> typeof(asset)
Asset{:BTC, :USDT}
end

Instruments.BaseCurrency — Type

A symbol checked to be a valid base currency.

Instruments.Cash — Type

A variable quantity of some currency.

> ca = c"USDT"
> typeof(ca)
# Instruments.Cash{:USDT}

Instruments.QuoteCurrency — Type

A symbol checked to be a valid quote currency.

Instruments.Derivatives.Derivative8 — Type

Derivative parsed accordingly to regex.

asset: The underlying asset.

sc: settlement currency.

id: identifier of the contract (the date).

strike: strike price.

kind: Instruments.Derivatives.DerivativeKind

Instruments.Derivatives.DerivativeKind — Type

Differentiates between perpetuals and options.

Instruments.Derivatives.SettlementCurrency — Type

A symbol parsed as settlement currency.

Instruments.deleverage_pair — Method

Remove leveraged pair pre/suffixes from base currency.

Instruments.isfiatpair — Method

Check if both base and quote are fiat currencies.

Instruments.islegeragedpair — Method

Test if pair has leveraged naming.

Instruments.subzero! — Method

Never subtract below zero.

Instruments.Derivatives.is_settled — Method

Predicates according to OctoBot

Instruments.@a_str — Macro

Parses pair to an Asset type. ```julia

typeof(a"BTC/USDT")

Instruments.Asset

Instruments.@c_str — Macro

Macro to instantiate Cash statically.

Don't put spaces between the id and the value.

> ca = c"USDT"1000
USDT: 1000.0

Instruments.@pusdt_str — Macro

Rewrites sym as a perpetual usdt symbol.

> pusdt"btc"
BTC/USDT:USDT

Instruments.FULL_SYMBOL_GROUPS_REGEX — Constant

From CCTX

Instruments.fiatnames — Constant

A list of fiat and fiat-like assets names.

Instruments.Derivatives.@d_str — Macro

Shortand for parsing derivatives:

> drv = d"BTC/USDT:USDT"
> typeof(drv)
# Instruments.Derivatives.Derivative{Asset{:BTC, :USDT}}