CoinMarketCap

Watchers.CoinMarketCap.credits — Method

Fetch remaining credits.

Watchers.CoinMarketCap.listings — Function

Fetch all coin listings.

Watchers.CoinMarketCap.quotes — Method

Call quotes to get data for a specific list of currencies.

Passing values of type Symbol will use the symbol parameter, while String will use the slug parameter.

Watchers.CoinMarketCap.setapikey! — Function

Sets coinmarketcap api key.

from env var CMC_API_KEY

or from config key coinmarketcap_apikey