Watchers

A Watcher type is an interface over a data feed. There are implementations for some 3d party apis, and exchange OHLCV data construction and order book.

User interface

To instantiate a watcher call its related function for example ccxt_ohlcv_watcher to instantiate a watcher that tracks trades data from an exchange and builds OHLCV candles for the specified timeframe. On a watcher instance these function are currently available:

get: What you mostly use watchers for. Get the underlying data monitored by the watcher, usually in a processed state (like a DataFrame), it defaults to the watcher buffer (which should keep data in a raw state).

length: Length of the underlying buffer.

last: Last raw value of the underlying buffer.

close: Stops the watcher and flushes the buffer.

isstale: Should evaluate if the watcher is in a degraded state, e.g. when it can't fetch new data.

fetch!: A watcher runs queries on specified interval, so you should only use fetch! when you want to be sure that the watcher has the latest data.

flush!: like fetch! the watcher already flushes at predetermined intervals, use this only to ensure flushing in case of shutdown. The watcher does call flush on destruction through its finalizer, but it does so asynchronously and doesn't ensure the success of the flush operation.

delete!: deletes the watcher data from the storage backend used by flush! (and empties the buffer).

deleteat!: deletes the watcher data within a date range (and empties the buffer).

Imlementation interface

To implement a custom watcher you have to define the functions such that dispatch happens through the watcher name interpreted as a value Val{Symbol(my_watcher_name)}. So a function needs to have a signature like _fetch!(w::Watcher, ::Val{some_symbol}).

Required

_fetch! whatever fetches the data, like an http request.

_get returns the post processed data, like a DataFrame.

Optional

_init! to perform initialization routines.

_load! to pre-fill the watcher buffer on construction, it is only called once, runs after _init!.

_flush! to save the watcher buffer somewhere on periodic intervals and on watcher destruction.

_process! to update the view of the raw data, which is what the get function should return.

_delete! to delete all the storage data of the watcher

_deleteat! to delete the storage data of the watcher within a date range (from, to)

Look in the Watchers and WatchersImpls modules for helper functions:

Watchers.Watchers — Module

Watchers are data feeds, that keep track of stale data.

Watchers.Watcher — Type

Watchers manage data, they pull from somewhere, keep a cache in memory, and optionally flush periodically to persistent storage.

buffer: A CircularBuffer of default length 1000 of the watcher type parameter.

name: The name is used for dispatching.

timer: A Timer, handles calling the function that fetches the data.

timeout: How much time to wait for the fetcher function.

interval: the Period with which the fetcher function will be called.

flush_interval: the Period with which the flusher function will be called.

attempts: In cause of fetcher failure, tracks how many consecutive fails have occurred. It resets after a successful fetch operation.

last_fetch: the most recent time a fetch operation failed.

last_flush: the most recent time the flush function was called.

Base.delete! — Method

Deletes all watcher data from storage backend. Also empties the buffer.

Base.deleteat! — Method

Delete watcher data from storage backend within the date range specified.

Base.get — Method

Returns the processed view of the watcher data. Accessible also as a view property of the watcher object.

Watchers._fetch! — Method

Appends new data to the watcher buffer, returns true when new data is added, false otherwise.

Watchers._flush! — Method

May run after a successful fetch operation, according to the flush_interval. It spawns a task.

Watchers._get — Method

Returns the processed view of the watcher data.

Watchers._init! — Method

Function to run on watcher initialization, it runs before _load!.

Watchers._load! — Method

Called once on watcher creation, used to pre-fill the watcher buffer.

Watchers._process! — Method

Processes the watcher data, called everytime the watcher fetches new data.

Watchers.default_flusher — Method

Save function for watcher data, saves to the default DATA_PATH located lmdb instance using serialization.

Watchers.default_loader — Method

Load function for watcher data, loads from the default DATA_PATH located lmdb instance using serialization.

Watchers.default_process — Method

Processes the values of a watcher buffer into a dataframe. To be used with: default_init and default_get

Watchers.fetch! — Method

Fetches a new value from the watcher ignoring the timer. If reset is true the timer is reset and polling will resume after the watcher interval.

Watchers.flush! — Method

Flush the watcher. If wait is true, block until flush completes.

Watchers.isstale — Method

True if last available data entry is older than now() + fetch_interval + fetch_timeout.

Watchers.pushnew! — Method

Helper function to push a new value to the watcher buffer if it is different from the last one.

Watchers.pushstart! — Method

Helper function to push a vector of values to the watcher buffer.

Watchers.watcher — Method

Instantiate a watcher.

T: The type of the underlying CircularBuffer

len: length of the circular buffer.

fetcher: The input function that fetches data from somewhere, with signature (Watcher) -> Bool.

flusher: Optional function that is called once every len successful updates, (Watcher) -> nothing

starter: Optional function that is called on watcher startup to prefill the buffer with previous (most recent) data, (Watcher) -> nothing

asyncio vs threads

Both fetcher and flusher callbacks assume non-blocking asyncio like behaviour. If instead your functions require high computation, pass threads=true, you will have to ensure thread safety.

Warning

If using flushers, do not modify the input data (argument of the flusher callback), always make a copy.

Watchers.@acquire — Macro

Same as @lock but with acquire and release.

Watchers.WatchersImpls.ccxt_ohlcv_watcher — Function

Create a Watcher instance that tracks ohlcv for an exchange (ccxt).

Watchers.WatchersImpls.ccxt_tickers_watcher — Function

Create a Watcher instance that tracks all markets for an exchange (ccxt).

Watchers.WatchersImpls.cg_derivatives_watcher — Method

Create a Watcher instance that tracks all the derivatives from an exchange.

Watchers.WatchersImpls.cg_ticker_watcher — Method

Create a Watcher instance that tracks the price of some currencies on an exchange (coingecko).

Watchers.WatchersImpls.cp_markets_watcher — Function

Create a Watcher instance that tracks all markets for an exchange (coinpaprika).

Watchers.WatchersImpls.cp_twitter_watcher — Function

Create a Watcher instance that tracks all markets for an exchange (coinpaprika).

Watchers.WatchersImpls.@append_dict_data — Macro

Defines a closure that appends new data on each symbol dataframe.

Watchers.WatchersImpls.@fromdict — Macro

This macro tries to fill a known NamedTuple from an unknown Dict.