Analysis

This module is pretty much a playground of functions that I am using/used, most likely of very little interest.

Analysis.corsp — Method

Compute spearman correlation between x and y, where y is shifted forward (lead).

Analysis.find_bottomed — Method

Find bottomed longs.

Analysis.find_peaked — Method

Find peaked shorts.

Analysis.fltsummary — Method

Return the summary of a filtered vector of pairdata.

Analysis.isrenkoentry — Method

A good renko entry is determined by X candles of the opposite color after Y candles.

Analysis.mark! — Method

Loads the Mark module.

Analysis.supres — Method

This support and resistance functions from Indicators appear to be too inaccurate despite parametrization.

Analysis.up_successrate — Method

Calculate successrate of given column against next candle. direction: true is buy, false is sell.

Base.filter — Method

Filters a list of pairs using a predicate function. The predicate functions must return a Real number which will be used for sorting.

Mark

MVP

MVP.MVP — Module

Momentum, volume, price indicator.

MVP.discrete_mvp — Method

MVP.is_mvp — Method

Returns the mvp-ness of a pair as a sum of each condition ratio weighted by weights. If real=false it will return a Bool indicating if the pair passes the given ratios.

MVP.momentum — Method

Ratio is the minimum number of green candles required.

MVP.volume — Method

Ratio is the minimum percent increment of volume from the first half of the series.

Short

Short.Short — Module

Predicates that signal lowered chances of success.

Short.find — Method

A good short should have high volume, (too) high price change, and dominating red candles.

Short.fullret — Method

Has a full retracement occurred?

Short.isbelow20 — Method

Is the close below the Average?

Short.isbelow50 — Method

Is the close below the average with volume above average? Intended to be used on longer windows in contrast to the volume-less version.

Short.isdcandles — Method

Are down candles more than up candles?

Short.lowhin — Method

True if series experienced a recent breakout with high volume retracement (Low out, High in). If breakout is not met, returns nothing. br: Breakout rule. window: How far backward to look for breakouts. delay: Delay to look for retracements.

Short.mustd — Method

Breakout level. Mean with std.

Long

Long.Long — Module

Predicates that signal increased chances of success.

Long.rs_cache — Constant

Since relative strength computes the score for the whole set, it is useful to cache it when querying elements from the same set.

Long.fthrough — Method

Follow through price action.

min_follow: Minimum number of upward candles that have to follow after breakout.

Long.istennisball — Method

Tennis ball action, resilient price snapback after a pullback.

snapback: The number of candles to consider for snapback action.

Long.long — Method

Evaluate traits for a collection of pairs.

Long.long — Method

Evaluate traits on multiple timeframes.

Long.long — Method

Evaluate trais for a single pair.

Long.pair_matrix — Method

Given a dict of dataframes (ohlcv) returns a tuple of matrices, where high and volume are respectively concatenated.

Long.relative_strength — Method

Computes the relative strength of a collection (dict) of pairs. Returning the queried one, and the rest.

sorted: Sort the matrix of all relative strengths.

norm_roc: Smooths the ranking by normalizing the rate of change at each candle.

Long.relative_strength — Method

Given a collection of price timeseries (1D), calculate the relative strength of each series as the summation of the rate of change at each timestep t.

Long.stage2 — Method

Stage 2 template.

base° is the maximum angle allowed for a pattern to be considered a base (above which would be uptrending).

baseσ is the window considered at the last step, for regressing the base formation.