Data

The main backend currently is Zarr. Zarr is similar to feather or parquet in that it optimizes to columnar data, or in general arrays. However it is simpler, and allows to pick different encoding schemes, and supports compression by default. More over the zarr interface can be backed by different storage layers, that can also be over the network. Compared to no/sql databases columnar storage has the drawback of having to read chunks for queries, but we are almost never are interested in scalar values, we always query a time-series of some sort, so the latency loss is a non issue.

Scraping

There are different ways to collect data:

Using the Scrapers module

Currently there is support for binance and bybit archives.

using Scrapers: Scrapers as scr, BinanceData as bn
## Download klines for ETH
bn.binancedownload("eth", market=:data, freq=:monthly, kind=:klines)
## load them
bn.binanceload("eth", market=:data, freq=:monthly, kind=:klines)
## Default market parameter is `:um` (usdm futures)

# show all symbols that can be downloaded
bn.binancesyms(market=:data)
# load/download also accept `quote_currency` to filter by (default `usdt`)
scr.selectsyms(["eth"], bn.binancesyms(market=:data), quote_currency="usdc")

!! "Downloads are cached" downloading the same pair path again will only downloads newer archives if data gets corrupted pass reset=true to redownload it again.

Using the Fetch module

The Fetch module downloads data directly from the exchange using ccxt.

using TimeTicks
using Exchanges
using Fetch: Fetch as fe

exc = getexchange!(:kucoin)
timeframe = tf"1m"
pairs = ("BTC/USDT", "ETH/USDT")
# Will fetch the last 1000 candles, `to` can also be passed to download a specific range
fe.fetch_candles(exc, timeframe, pairs; from=-1000)

Fetching directly from exchanges is not recommended for smaller timeframes since they are heavily rate limited. Archives are better.

Using Watchers

With the Watchers module you can track live data from exchanges or other data sources and store it locally. Implemented are watchers that track OHLCV:

using Exchanges
using PingPong.Watchers: Watchers as wc, WatchersImpls as wi
exc = getexchange!(:kucoin)

w = wi.ccxt_ohlcv_tickers_watcher(exc;)
wc.start!(w)

>>> w
17-element Watchers.Watcher20{Dict{String, NamedTup...Nothing, Float64}, Vararg{Float64, 7}}}}}
Name: ccxt_ohlcv_ticker
Intervals: 5 seconds(TO), 5 seconds(FE), 6 minutes(FL)
Fetched: 2023-03-07T12:06:18.690 busy: true
Flushed: 2023-03-07T12:04:31.472
Active: true
Attemps: 0

As a convention the view property of a watcher shows the processed data. In this case the candles processed by the ohlcv_ticker_watcher will be stored in a dict.

>>> w.view
Dict{String, DataFrames.DataFrame} with 220 entries:
  "HOOK/USDT"          => 5×6 DataFrame…
  "ETH/USD:USDC"       => 5×6 DataFrame…
  "PEOPLE/USDT:USDT"   => 5×6 DataFrame…

There is another OHLCV watcher based on trades, that tracks only one pair at a time.

w = wi.ccxt_ohlcv_watcher(exc, "BTC/USDT:USDT"; timeframe=tf"1m")
w.view
956×6 DataFrame
 Row │ timestamp            open     high     low      close    volume  
     │ DateTime             Float64  Float64  Float64  Float64  Float64 
─────┼──────────────────────────────────────────────────────────────────
...

Other implemented watchers are the orderbook watcher, and watchers that parse data feeds from 3rd party apis.

API

Data.OHLCVTuple — Type

Similar to a StructArray (and should probably be replaced by it), used for fast conversion.

Data.ZarrInstance — Type

Candles data is stored with hierarchy PAIR -> [TIMEFRAMES...]. A pair is a ZGroup, a timeframe is a ZArray.

Base.unique! — Method

Remove duplicate from a zarray.

In a 2d zarray where we want values where the second column is unique:

unique!(x->x[2], z)

Data._load_ohlcv — Method

Load ohlcv pair data from zarr instance. za: The zarr array holding the data key: the name of the array to load from the zarr instance (in the format exchange/timeframe/pair) td: the timeframe (as integer in milliseconds) of the target ohlcv table to be loaded from, to: date range

Data._save_data — Method

Saves data to a zarr array ensuring only dates seriality, not contiguity (as opposed to _save_ohlcv).

Data.candleat — Method

Get the candle at given date from a ohlcv dataframe as a Candle.

Data.candlelast — Method

Get the last candle from a ohlcv dataframe as a Candle.

Data.candlepair — Method

Same as candleat but also fetches the previous candle, returning a Tuple{Candle, Candle}.

Data.chunksize — Method

Choose chunk size depending on size of data with a predefined split (e.g. 1/100), padding to the nearest power of 2.

Data.contiguous_ts — Method

Checks if a timeseries has any intervals not conforming to the given timeframe.

Data.df! — Method

Construct a DataFrame without copying.

Data.empty_ohlcv — Method

An empty OHLCV dataframe.

Data.key_path — Method

The full key of the data stored for the (exchange, pair, timeframe) combination.

Data.load — Method

Load a pair ohlcv data from storage. as_z: returns the ZArray

Data.load_data — Method

Load data from zarr instance. zi: The zarr instance to use key: the name of the array to load from the zarr instance (full key path). type: Set to the type that zarr should use to store the data (only bits types). [Float64]. serialized: If set, data will be deserialized before returned (type is ignored). from, to: date range sz: The chunks tuple which should match the shape of the already saved data.

Mismatching chunks

Loading data with from key with wrong dimensions (ndims(sz)) or shape (columns) will reset the stored zarray.

Data.load_ohlcv — Method

Load data from given zarr instance, exchange, pairs list and timeframe.

Data.save_ohlcv — Method

data_col: the timestamp column of the new data (1) saved_col: the timestamp column of the existing data (1) pair: the trading pair (BASE/QUOTE string) timeframe: exchange timeframe (from exc.timeframes) type: Primitive type used for storing the data (Float64) check:

:bounds (default) only checks that new data is adjacent to previous data.

:all checks full contiguity of previous and new data.

:none or anything else, no checks are done.

Data.snakecased — Method

Normalizes or special characthers separators to _.

Data.to_ohlcv — Method

Convert raw ccxt OHLCV data (matrix) to a dataframe.

Data.zdelete! — Method

Delete elements from a ZArray within the range of dates from_dt:to_dt.

Use the select function to customize the dates vector where to index th range. Defaults to the first column of a 2D Zarray.

Data.zilmdb — Function

Create a ZarrInstance at specified path using lmdb as backend.

force: resets the underlying store.

Data.@as_mat — Macro

Redefines given variable to a Matrix with type of the underlying container type.

Data.@as_ts — Macro

Converts a subset of columns to timeseries.

Data.@ohlc — Macro

Converts ohlcv dataframe to timeseries type.

Data.@otime — Macro

Get the last date of loaded data (in mrkts variable).

Data.@to_mat — Macro

Same as as_mat but returns the new matrix.

Prices

Main.Prices.gprofit — Function

Total profit of a ladder of trades

Main.Prices.price_ranges — Function

Given a price, output price at given ratios. (Predefined ratios - :bal,:futures,:tight,:expa)

Main.Prices.price_ranges — Method

Get the price range of a map of pairs, using the last available close price.