Engine

Backtesting is based on this assumptions

Engine.Context8 — Type

The configuration against which a strategy is tested.

timeframe: The time step or "tick" that the backtesting iteration uses.

spread: affects the weight of the spread calculation (based on ohlcv)."

slippage: affects the weight of the spread calculation (based on volume and trade size).

Engine.backtest! — Method

Backtest passing a context arguments as a tuple.

Engine.backtest! — Method

Backtest a strategy strat using context ctx iterating according to the specified timeframe.

On every iteration, the strategy is queried for the current timestamp. The strategy should only access data up to this point. Example:

Timeframe iteration: 5s

Strategy minimum available timeframe 1m

Iteration gives time 1999-12-31T23:59:59 to the strategy: The strategy (that can only lookup up to 1m precision) looks-up data until the timestamp 1999-12-31T23:58:00 which represents the time until 23:59:00. Therefore we have to shift by one period down, the timestamp returned by apply:

julia> t = TimeTicks.apply(tf"1m", dt"1999-12-31T23:59:59")
1999-12-31T23:59:00 # we should not access this timestamp
julia> t - tf"1m".period
1999-12-31T23:58:00 # this is the correct candle timestamp that we can access

To avoid this mistake, use the function available(::TimeFrame, ::DateTime), instead of apply.

Engine.Instances.AssetInstance27 — Type

An asset instance holds all known state about an asset, i.e. BTC/USDT:

asset: the identifier

data: ohlcv series

history: the trade history of the pair

cash: how much is currently held, can be positive or negative (short)

exchange: the exchange instance that this asset instance belongs to.

limits: minimum order size (from exchange)

precision: number of decimal points (from exchange)

Base.fill! — Method

Pulls data from storage, or resample from the shortest timeframe available.

Engine.Instances.last_candle — Method

Get the last available candle strictly lower than apply(tf, date)

Engine.Instances.load! — Method

Load ohlcv data of asset instance.

Engine.Collections.AssetCollection2 — Type

A collection of assets instances, indexed by asset and exchange identifiers.

Base.fill! — Method

Main.Engine.Instances.fill! all the instances with given timeframes data...

Engine.Collections.bqe — Method

Dispatch based on either base, quote currency, or exchange.

Engine.Collections.flatten — Method

Returns a Dict{TimeFrame, DataFrame} of all the OHLCV dataframes present in the asset collection.