Engine

Backtesting

It is based on some assumptions

API

Engine.Types.Instances.AssetInstance49 — Type

An asset instance holds all known state about an asset, i.e. BTC/USDT:

asset: the identifier

data: ohlcv series

history: the trade history of the pair

cash: how much is currently held, can be positive or negative (short)

exchange: the exchange instance that this asset instance belongs to.

limits: minimum order size (from exchange)

precision: number of decimal points (from exchange)

Base.fill! — Method

Pulls data from storage, or resample from the shortest timeframe available.

Base.similar — Method

Returns a similar asset instance with cash and orders reset.

Data.candlelast — Method

Get the last available candle strictly lower than apply(tf, date)

Engine.Types.Instances.load! — Method

Load ohlcv data of asset instance.

Engine.Types.Collections.AssetCollection2 — Type

A collection of assets instances, indexed by asset and exchange identifiers.

TimeTicks.DateRange — Type

Makes a daterange that spans the common min and max dates of the collection.

Base.fill! — Method

Main.Engine.Types.Instances.fill! all the instances with given timeframes data...

Data.stub! — Method

Replaces the data of the asset instances with src which should be a mapping. Used for backtesting.

src: The mapping, should be a pair TimeFrame => Dict{String, PairData}.

Example:

using Scrapers.BinanceData as bn
using Strategies
using Exchanges
setexchange!(:binanceusdm)
cfg = Config(nameof(exc.id))
strat = strategy!(:Example, cfg)
data = bn.binanceload()
stub!(strat.universe, data)

Engine.Types.Collections.bqe — Method

Dispatch based on either base, quote currency, or exchange.

Engine.Types.Collections.flatten — Method

Returns a Dict{TimeFrame, DataFrame} of all the OHLCV dataframes present in the asset collection.

Engine.Types.Collections.iscashable — Method

Checks that all assets in the universe match the cash.