Exchanges

Exchange Types

Basic exchange types, and global exchange vars.

ExchangeTypes.exc — Constant

Global var implicit exchange instance.

When working interactively, a global exc variable is available, updated through globalexchange!, which is used as the default for some functions when the exchange argument is omitted.

ExchangeTypes.exchanges — Constant

Global var holding Exchange instances. Used as a cache.

ExchangeTypes.ExcPrecisionMode — Type

Same as ccxt precision mode enums.

ExchangeTypes.Exchange — Type

The exchange type wraps a ccxt exchange instance. Some attributes frequently accessed are copied over to avoid round tripping python. More attributes might be added in the future. To instantiate an exchange call getexchange! or setexchange!.

ExchangeTypes.ExchangeID — Type

An ExchangeID is a symbol checked to match a ccxt exchange class.

Base.getproperty — Method

Attributes not matching the Exchange struct fields are forwarded to the wrapped ccxt class instance.

Base.hash — Method

The hash of an exchange object is reduced to its symbol (the function used to instantiate the object from ccxt).

ExchangeTypes.globalexchange! — Method

Updates the global exchange exc variable.

Exchange functions

Exchanges.exckeys! — Method

Set exchange api keys.

Exchanges.futures — Method

Returns the matching futures exchange instance, if it exists, or the input exchange otherwise.

Exchanges.get_markets — Method

Get the the markets of the ccxt instance, according to min_volume and quote currency.

Exchanges.getexchange! — Method

getexchage!: ccxt exchange by symbol either from cache or anew. It uses a WS instance if available, otherwise an async instance.

Exchanges.hastickers — Method

Check if exchange has tickers list.

Exchanges.issandbox — Function

Check if sandbox mode is enabled for exchange.

Exchanges.loadmarkets! — Method

Load exchange markets:

cache: rely on storage cache

agemax: max cache valid period [1 day].

Exchanges.price_ranges — Method

Get price ranges using tickers data from exchange.

Exchanges.quotevol — Method

Get quote volume from ticker.

Exchanges.ratelimit! — Function

Enable or disable rate limit.

Exchanges.sandbox! — Function

Enable sandbox mode for exchange

Exchanges.setexchange! — Method

Instantiate an exchange struct. it sets:

The matching ccxt class.

Pre-emptively loads the markets.

Sets the exchange timeframes.

Sets exchange api keys.

Exchanges.spotsymbol — Method

Trims the settlement currency in futures. (mkt is a ccxt market.)

Exchanges.tickerprice — Method

Get price from ticker.

Exchanges.timeout! — Function

Set exchange timouet

Exchanges.@exchange! — Macro

Define an exchange variable set to its matching exchange instance.

Exchanges.@tickers — Macro

Fetch and cache tickers data.

Exchanges.is_pair_active — Function

Exchanges.market! — Function

Retrieves a cached market (1minute) or fetches it from exchange.

Exchanges.market_fees — Function

Taker fees for market.

Exchanges.market_limits — Function

Minimum order size of the of the market.

Exchanges.market_precision — Function

Precision of the (base, quote) currencies of the market.

Exchanges.tickers — Method

Get the exchange tickers.

quot: Only choose pairs where the quot currency equals quot.

min_vol: The minimum volume of each pair.

skip_fiat: Ignore fiat/fiat pairs.

margin: Only choose pairs enabled for margin trading.

leveraged: If :no skip all pairs where the base currency matches the leverage_pair_rgx regex.

as_vec: Returns the pairlist as a Vector instead of as a Dict.

Data

Extensions for the top level Data module.

Data.load_ohlcv — Method

Load all pairs from exchange according to config quote currency and timeframe.

Data.save_ohlcv — Method

Updates pair data of the globally set exchange.

Fetching

Helper module for downloading data off exchanges.

Fetch.SINCE_INC — Constant

Used to slide the since param forward when retrying fetching (in case the requested timestamp is too old).

Base.convert — Method

This is the fastest (afaik) way to convert ccxt lists to dataframe friendly format.

Fetch._fetch_loop — Method

Calls the fetc_func iteratively until the full dates range has been downloaded. NOTE: The returned data won't be exactly the number of candles expected by e.g. length(DateRange(from, to))

Fetch._fetch_ohlcv_from_to — Method

Ensure a to date is set, before fetching.

Fetch._fetch_with_delay — Method

Wraps a fetching function around error handling and backoff delay. fetch_func signature: (pair::String, since::Option{Float}, limit::Float) -> PyList The converter function has to tabulate the data such that the first column is the timestamp

Fetch.fetch_funding — Method

Fetch funding rate history from exchange for a list of Derivative pairs.

from, to: specify date period to fetch candles for.

Fetch.fetch_ohlcv — Method

Fetch ohlcv data from exchange for a list of pairs.

from, to: Can represent a date. A negative from number implies fetching the last N=from candles.

update: If true, will check for cached data, and fetch only missing candles. (false)

progress: if true, show a progress bar. (true)

reset: if true, will remove cached data before fetching. (false)

Fetch.fetch_ohlcv — Method

Fetch ohlcv data for multiple exchanges on the same timeframe.

It accepts:

a mapping of exchange instances to pairlists.

a vector of symbols for which an exchange instance will be instantiated for each element, and pairlist will be composed according to quote currency and min_volume from PingPong.config.

Fetch.find_since — Method

Should return the oldest possible timestamp for a pair, or something close to it.

Fetch.funding_data — Method

Retrieves all funding data return by exchange for symbol, or a subset.

funding_data(exc, "BTC/USDT:USDT")
funding_data(exc, "BTC/USDT:USDT", :fundingRate, :markPrice)