Misc

Misc.config — Constant

Global configuration instance.

Misc.futures_exchange — Constant

Some exchanges are split into different classes in ccxt.

Misc.ohlcv_limits — Constant

Exchange ohlcv candles limits.

Misc.results — Constant

Holds recently evaluated statements.

Misc.Config12 — Type

The config main structure:

window: The default number of candles (OHLCV).

timeframe: The default timeframe of the candles.

qc: The default quote currency.

margin: If margin is enabled, only margin pairs are considered.

leverage:

:yes : leveraged pairs will not be filtered.

:only : ONLY leveraged will not be filtered.

:from : Selects non leveraged pairs, that also have a leveraged siblings.

futures: Selects the futures version of an Exchange.

attrs: Generic metadata container.

sources: mapping of modules symbols name to (.jl) file paths

Misc._instantiate_workers — Method

Instantiate new workers if the current number mismatches the requested one.

Misc.config_path — Method

The config path (TOML), relative to the current project directory.

Misc.loadconfig! — Method

Parses the toml file and populates the global config.

Misc.resetconfig! — Method

Reset global config to default values.

Misc.setcfg! — Method

Sets a single config value.

Misc.@as_dfdict — Macro

Binds a mrkts variable to a Dict{String, DataFrame} where the keys are the pairs names and the data is the OHLCV data of the pair.

Misc.@lev! — Macro

Toggle config leverage flag

Misc.@margin! — Macro

Toggle config margin flag.

Progress bar

Pbar.Pbar — Module

Python enlighten progressbar wrapper.

Pbar.pbclose — Method

Terminates the progress bar.

Pbar.@pbar! — Macro

Instantiate a progress bar:

data: length(data) determines the bar total

unit: what unit the display

desc: description will appear over the progressbar

Pbar.@pbupdate! — Macro

Single update to the progressbar with the new value.