Processing

Processing.cleanup_ohlcv_data — Method

Similar to the freqtrade homonymous function.

fill_missing: :close fills non present candles with previous close and 0 volume, else with NaN.

Processing.fill_missing_rows — Method

Assuming timestamps are sorted, returns a new dataframe with a contiguous rows based on timeframe. Rows are filled either by previous close, or NaN.

Processing.is_incomplete_candle — Method

Checks if a candle timestamp is too new.

Processing.is_last_complete_candle — Method

Checks if a timestamp belongs to the newest possible candle of given timeframe.

Processing.maptf — Method

Apply a function over data, resampling data to each timeframe in tfs.

f: signature is (data; kwargs...)::DataFrame

tfsum: sum the scores across multiple timeframes for every pair.

Processing.normalize! — Method

Applies either a unitrange transform or a zscore tranform over the data in place.

Processing.resample — Method

Resamples ohlcv data from a smaller to a higher timeframe.