Strategy interface

Load a strategy

The strategy is instantiated by loading a julia module at runtime.

using PingPong
cfg = Config(:kucoin) # Load the configuration, choosing kucoin as exchange
strategy!(:Example, cfg) # Load the Example strategy

The strategy is looked up inside the config under the sources key:

# Example config
[kucoin]
futures = true
[sources]
Example = "cfg/strategies/Example.jl" # the name of the module

The key is the name of the module (in this case Example) which will be imported from the included file "cfg/strategies/Example.jl".

After the strategy module is imported the strategy is instantiated by calling the ping!(::Type{S}, ::LoadStrategy, cfg) function.

> typeof(s)
Engine.Strategies.Strategy37{:Example, ExchangeTypes.ExchangeID{:kucoin}(), :USDT}

See here how the load method is defined.

module Example
using Engine.Strategies
using ExchangeTypes

const NAME = :Example
const EXCID = ExchangeID(:bybit)
const S{M} = Strategy{M,NAME,typeof(EXCID)}
const TF = tf"1m"

function ping!(::Type{S}, ::LoadStrategy, config)
    assets = marketsid(S)
    s = Strategy(Example, assets; config)
    s
end

end

See that the load method dispatches on the strategy type with cfg as argument of type Misc.Config.

As a rule of thumb if the method should be called before the strategy is construct, then it dispatches to the strategy type (Type{S}), otherwise the strategy instance (S). For convention the module property S of your strategy module, declares the strategy type (const S = Strategy{name, exc}).

Strategy interface

ping!(::Type{S}, ::LoadStrategy, config): loads the strategy

ping!(::S, ::WarmupPeriod): returns how much data the strategy needs on startup

ping!(::S, ::DateTime, ::Context): the "main" strategy function, called once per candle in backtest and once per throttle during live.

API

Engine.Strategies.Strategy — Type

The strategy is the core type of the framework.

The strategy type is concrete according to:

Name (Symbol)

Exchange (ExchangeID), read from config

Quote cash (Symbol), read from config

The exchange and the quote cash should be specified from the config, or the strategy module.

universe: All the assets that the strategy knows about

holdings: assets with non zero balance.

orders: active orders

timeframe: the smallest timeframe the strategy uses

cash: the quote currency used for trades

Conventions for strategy defined attributes:

NAME: the name of the strategy could be different from module name

S: the strategy type.

TF: the smallest timeframe that the strategy uses

Engine.Types.Context — Method

Creates a context within the available data loaded into the strategy universe with the smallest timeframe available.

Engine.Strategies.assets — Method

Assets loaded by the strategy.

Engine.Strategies.default! — Method

Set strategy defaults.

Engine.Strategies.exchange — Method

Strategy main exchange id.

Engine.Strategies.freecash — Method

Cash that is not committed, and therefore free to use for new orders.

Engine.Strategies.instances — Method

Strategy assets instance.

Engine.Strategies.ping! — Method

Called on each timestep iteration, possible multiple times. Receives:

current_time: the current timestamp to evaluate (the current candle would be current_time - timeframe).

ctx: The context of the executor.

Engine.Strategies.ping! — Method

How much lookback data the strategy needs.

Engine.Strategies.ping! — Method

Called to construct the strategy, should return the strategy instance.

Engine.Strategies.reload! — Method

Reloads ohlcv data for assets already present in the strategy universe.

Engine.Strategies.reset! — Method

Resets strategy.

Engine.Strategies.tradesedge — Method

Returns the first and last trade of any asset in the strategy universe.

Misc.execmode — Method

Returns the strategy execution mode.