Strategy interface

Load a strategy

The strategy is instantiated by loading a julia module at runtime.

using PingPong
cfg = loadconfig!(:kucoin, cfg=Config()) # Load the configuration, choosing kucoin as exchange
loadstrategy!(:Macd, cfg) # Load the Macd strategy

The strategy is looked up inside the config under the sources key:

# Example config
[kucoin]
futures = true
[sources]
Macd = "cfg/strategies/Macd.jl" # the name of the module

The key is the name of the module (in this case Macd) which will be imported from the included file "cfg/strategies/Macd.jl".

After the strategy module is imported the strategy is instantiated according to the module name (the strategy type is parametric).

s = Strategy(name)

> typeof(s)
Engine.Strategies.Strategy37{:Macd, ExchangeTypes.ExchangeID{:kucoin}(), :USDT}

Familiarize with the data structures

Engine.Strategies.Strategy — Type

The strategy is the core type of the framework.

The strategy type is concrete according to:

Name (Symbol)

Exchange (ExchangeID), read from config

Quote cash (Symbol), read from config

The exchange and the quote cash should be specified from the config, or the strategy module.

universe: All the assets that the strategy knows about

balances: assets with open orders or non zero balance.

orders: all active orders

cash: the quote currency used for trades

Engine.Strategies.clearorders! — Method

Clears all orders history from strategy.

Engine.Strategies.reload! — Method

Reloads ohlcv data for assets already present in the strategy universe.