Strategy interface

Setup a new strategy

The simplest way to create a strategy is to use the interactive generator which will prompt for the required set of options to set.

julia> using PingPong
julia> PingPong.generate_strategy()
Strategy name: : MyNewStrategy

Timeframe:
   1m
 > 5m
   15m
   1h
   1d

Select exchange by:
 > volume
   markets
   nokyc

 > binance
   bitforex
   okx
   xt
   coinbase

Quote currency:
   USDT
   USDC
 > BTC
   ETH
   DOGE

Margin mode:
 > NoMargin
   Isolated

Activate strategy project at /run/media/fra/stateful-1/dev/PingPong.jl/user/strategies/MyNewStrategy? [y]/n: y

Add project dependencies (comma separated): Indicators
   Resolving package versions...
   [...]
  Activating project at `/run/media/fra/stateful-1/dev/PingPong.jl/user/strategies/MyNewStrategy`

┌ Info: New Strategy
│   name = "MyNewStrategy"
│   exchange = :binance
└   timeframe = "5m"
[ Info: Config file updated

Load strategy? [y]/n: 

julia> s = ans

Alternatively you can directly pass kwargs and skip interaction by passing ask=false.

PingPong.generate_strat("MyNewStrategy", ask=false, exchange=:myexc)

or just use a config:

cfg = PingPong.Config(exchange=:myexc)
PingPong.generate_strat("MyNewStrategy", cfg)

Load a strategy

The strategy is instantiated by loading a julia module at runtime.

using PingPong
cfg = Config(exchange=:kucoin) # Constructs a configuration object, choosing kucoin as exchange
s = strategy(:Example, cfg) # Load the Example strategy

The key is the name of the module (in this case Example) which will be imported from the included file "cfg/strategies/Example.jl" or "cfg/strategies/Example/src/Example.jl".

After the strategy module is imported the strategy is instantiated by calling the ping!(::Type{S}, ::LoadStrategy, cfg) function.

> typeof(s)
Engine.Strategies.Strategy37{:Example, ExchangeTypes.ExchangeID{:kucoin}(), :USDT}

See here how the load method is defined.

module Example
using PingPong

const DESCRIPTION = "Example"
const EXC = :phemex
const MARGIN = NoMargin
const TF = tf"1m"

@strategyenv!

function ping!(::Type{<:SC}, ::LoadStrategy, config)
    assets = marketsid(S)
    s = Strategy(Example, assets; config)
    s
end

end

See that the load method dispatches on the strategy type with cfg as argument of type Misc.Config.

As a rule of thumb if the method should be called before the strategy is constructed, then it dispatches to the strategy type (Type{<:S}), otherwise the strategy instance (S). For convention the module property S of your strategy module, declares the strategy type (const S = Strategy{name, exc, ...}) and SC defines the same strategy type where the exchange is still generic.

Manual setup

If you want to create a strategy manually you can either:

Copy the user/strategies/Template.jl to a new file in the same directory and customize it.

Generate a new project in user/strategies and customize Template.jl to be your project entry file. The strategy Project.toml is used to store strategy config options. See other strategies examples for what the keys that are required.

For more advanced setups you can also use PingPong as a library, and construct the strategy object directly from your own module:

using PingPong
using MyDownStreamModule
s = PingPong.Engine.Strategies.strategy(MyDownStreamModule)

Strategy interface

Both ping! and pong! functions adhere to a convention for function signatures. The first argument is always either an instance of the subject or its type, followed by the arguments of the function, with the last non kw argument being the verb which describes the purpose of the function. KW arguments are optional and don't have any requirements. We can see below that Type{S} is the subject, config is an argument, and ::LoadStrategy is the _verb.

List of strategy ping! functions

Misc.ping! — Function

Called on each timestep iteration, possible multiple times. Receives:

current_time: the current timestamp to evaluate (the current candle would be current_time - timeframe).

ctx: The context of the executor.

ping!(
    _::Strategies.Strategy,
    current_time::Dates.DateTime,
    ctx
)

Called to construct the strategy, should return the strategy instance.

ping!(
    _::Type{<:Strategies.Strategy},
    cfg,
    _::Strategies.LoadStrategy
)

Called at the end of the reset! function applied to a strategy.

ping!(_::Strategies.Strategy, _::Strategies.ResetStrategy)

How much lookback data the strategy needs.

ping!(
    s::Strategies.Strategy,
    _::Strategies.WarmupPeriod
) -> Any

When an order is canceled the strategy is pinged with an order error.

ping!(
    s::Strategies.Strategy,
    ::OrderTypes.Order,
    err::OrderTypes.OrderError,
    ::Instances.AssetInstance;
    kwargs...
) -> Array{OrderTypes.StrategyEvent{E}, 1} where E<:ExchangeID

Market symbols that populate the strategy universe

After a position was updated from a trade.

ping!(
    _::Strategies.Strategy{X, N, <:ExchangeID, <:Misc.WithMargin, C} where {X<:Misc.ExecMode, N, C},
    ai,
    trade::OrderTypes.Trade,
    _::Instances.Position,
    _::Instances.PositionChange
)

This function is called after a position is updated due to a trade. It takes in a MarginStrategy, ai, trade, Position, and PositionChange as arguments. The function does not return any value.

After a position update from a candle.

ping!(
    _::Strategies.Strategy{X, N, <:ExchangeID, <:Misc.WithMargin, C} where {X<:Misc.ExecMode, N, C},
    ai,
    date::Dates.DateTime,
    _::Instances.Position,
    _::Instances.PositionUpdate
)

This function is called after a position is updated from a candle. It provides the necessary functionality for handling position updates in response to candle data.

Returns Optimizations.ContextSpace for backtesting

ping!(_::Strategies.Strategy, _::Executors.OptSetup)

The ctx field (Executors.Context) specifies the backtest time period, while space is either an already built BlackBoxOptim.SearchSpace subtype or a tuple (Symbol, args...) for a pre-defined BBO package search space.

Applies parameters to strategy before backtest

ping!(_::Strategies.Strategy, params, _::Executors.OptRun)

Removing a strategy

The function remove_strategy allows to discard a strategy by its name. It will delete the julia file or the project directory and optionally the config entry.

julia> PingPong.remove_strategy("MyNewStrategy")
Really delete strategy located at /run/media/fra/stateful-1/dev/PingPong.jl/user/strategies/MyNewStrategy? [n]/y: y
[ Info: Strategy removed
Remove user config entry MyNewStrategy? [n]/y: y

Strategy examples

Strategy examples can be found in the user/strategies folder, some strategies are single files like Example.jl while strategies like BollingerBands or ExampleMargin are project based.

Resizeable universe

The universe (s.universe) is backed by a DataFrame (s.universe.data). It is possible to add and remove assets from the universe during runtime, (although not extensively tested).