CoinGecko

API

Watchers.CoinGecko.coinschart — Method

Get close price and volume for symbol id for specified number of days.

days="max"

returns all available history.

Watchers.CoinGecko.coinschart_range — Method

Same as CoinGecko.coinschart but for specified timerange.

from, to: date range boundaries.

From coingecko:

Data granularity is automatic (cannot be adjusted)

1 day from current time = 5 minute interval data

1 - 90 days from current time = hourly data

above 90 days from current time = daily data (00:00 UTC)

Watchers.CoinGecko.coinschart_tf — Method

Pull market data range according to timeframe.

5m: 1 day

1h: 90 days

>1h: 365 days

Watchers.CoinGecko.coinshistory — Function

Get data for symbol id at specified date.

Watchers.CoinGecko.coinsid — Method

Get all current data for symbol id.

Watchers.CoinGecko.coinsmarkets — Method

Get markets for a list of symbols, accepting params CoinGecko.Params.

Watchers.CoinGecko.coinsohlc — Function

Pulls ohlc data (no volumes) according to timeframe.

<=30m: 1 day

<=4h: 30 days

>4h: max

Watchers.CoinGecko.coinsticker — Method

Get all tickers (from all coingecko exchanges) for symbol id.

Watchers.CoinGecko.derivatives — Method

Returns all unexpired derivative contracts.

Contracts are grouped by exchange, the id being the slugified exchange name.

Watchers.CoinGecko.derivatives_from — Method

Fetch derivatives from specified exchange.

Returns a Dict{Derivative, Dict}.

Watchers.CoinGecko.globaldata — Method

Returns global data for:

volume: total_volume (Dict{String, Float64})

mcap_change_24h: marketcapchangepercentage24h_usd

Watchers.CoinGecko.idbysym — Method

Get all coingecko item id matching by its symbol.

Watchers.CoinGecko.idbysym — Method

Get the first coingecko item id by its symbol.

Watchers.CoinGecko.loadcoins! — Method

Load all coins symbols.

Watchers.CoinGecko.loadderivatives! — Method

Returns the list of all exchange ids.

Watchers.CoinGecko.ratelimit — Method

Allows only 1 query every 3000 milliseconds seconds.

Watchers.CoinGecko.tickers_from — Method

Fetches top 100 tickers from exchange.

Returns a Dict{Asset, Dict}

Watchers.CoinGecko.trending — Method

24h trending top 7 coins.