CoinPaprika

API

Watchers.CoinPaprika.addcalls! — Function

Manually add available api calls (mostly for debugging).

Watchers.CoinPaprika.betas — Method

Returns coins betas in a dataframe compatible type.

Watchers.CoinPaprika.coin_exchanges — Method

Get all exchanges for specified coin.

Watchers.CoinPaprika.coin_markets — Method

Returns all markets for give coin (interpreted as base currency).

Expensive call

Don't call too often.

Watchers.CoinPaprika.coin_ohlcv — Method

Coin ohlcv (last day).

incomplete: if true returns the (today) incomplete candle.

Watchers.CoinPaprika.hourly — Function

Returns historical hourly tick values for the last day.

Watchers.CoinPaprika.idbysym — Method

Get all coinpaprika item id matching by its symbol.

Watchers.CoinPaprika.idbysym — Method

Get the first coinpaprika item id by its symbol.

Watchers.CoinPaprika.loadcoins! — Method

Load all coin ids.

Watchers.CoinPaprika.loadexchanges! — Method

Load market ids.

Watchers.CoinPaprika.markets — Function

Fetch all markets for specified exchange.

Watchers.CoinPaprika.ratelimit — Method

CoinPaprika free plan is 25k queries per month, which is 1q/25s.

On every call we check when the last query was performed, and add available queries that weren't used on a counter, to allow for bursts.

Not precise

Coinpaprika does not expose credits consumed by each endpoint, so we assume all calls are equals (2/min).

Watchers.CoinPaprika.ticker — Method

Fetch ticker for specified coin.

Watchers.CoinPaprika.tickers — Function

Fetch quotes for all pairs, includes coinpaprika beta metric.

Watchers.CoinPaprika.twitter — Method

Get last ~50 tweets for coin.