Watchers

A Watcher type is an interface over a data feed. There are implementations for some 3d party apis, and exchange OHLCV data construction and order book.

User interface

To instantiate a watcher call its related function for example ccxt_ohlcv_watcher to instantiate a watcher that tracks trades data from an exchange and builds OHLCV candles for the specified timeframe. On a watcher instance these function are currently available:

get: What you mostly use watchers for. Get the underlying data monitored by the watcher, usually in a processed state (like a DataFrame), it defaults to the watcher buffer (which should keep data in a raw state).

length: Length of the underlying buffer.

last: Last raw value of the underlying buffer.

close: Stops the watcher and flushes the buffer.

isstale: Should evaluate if the watcher is in a degraded state, e.g. when it can't fetch new data.

fetch!: A watcher runs queries on specified interval, so you should only use fetch! when you want to be sure that the watcher has the latest data.

flush!: like fetch! the watcher already flushes at predetermined intervals, use this only to ensure flushing in case of shutdown. The watcher does call flush on destruction through its finalizer, but it does so asynchronously and doesn't ensure the success of the flush operation.

delete!: deletes the watcher data from the storage backend used by flush! (and empties the buffer).

deleteat!: deletes the watcher data within a date range (and empties the buffer).

push!: add an element to the elements the watcher subscribes to (if any).

pop!: opposite of push!.

stop: Stops the watcher.

start: Restarts the watcher.

Imlementation interface

To implement a custom watcher you have to define the functions such that dispatch happens through the watcher name interpreted as a value Val{Symbol(my_watcher_name)}. So a function needs to have a signature like _fetch!(w::Watcher, ::Val{some_symbol}).

Required

_fetch! whatever fetches the data, like an http request.

_get returns the post processed data, like a DataFrame.

Optional

_init! to perform initialization routines.

_load! to pre-fill the watcher buffer on construction, it is only called once, runs after _init!.

_flush! to save the watcher buffer somewhere on periodic intervals and on watcher destruction.

_process! to update the view of the raw data, which is what the get function should return.

_delete! to delete all the storage data of the watcher.

_deleteat! to delete the storage data of the watcher within a date range (from, to).

_push! watchers might manage a list of things to track (like Asset symbols).

_pop! inverse of _push!.

_start Executed before starting the timer.

_stop Executed after stopping the timer.

Look in the Watchers and WatchersImpls modules for helper functions:

API

Watchers.Watchers — Module

Watchers are data feeds, that keep track of stale data.

Watchers.Watcher — Type

Watchers manage data, they pull from somewhere, keep a cache in memory, and optionally flush periodically to persistent storage.

buffer: A CircularBuffer of default length 1000 of the watcher type parameter.

name: The name is used for dispatching.

has: flags that show which callbacks are enabled between load, process and flush.

fetch_timeout: How much time to wait for the fetcher function.

fetch_interval: the Period with which _fetch! function will be called.

flush_interval: the Period with which _flush! function will be called.

capacity: controls the size of the buffer and the processed container.

beacon: Conditions notified on successful fetch, process and flush events.

threads: flag to enable to execute fetching in a separate thread.

attempts: In cause of fetching failure, tracks how many consecutive fails have occurred. It resets after a successful fetch operation.

last_fetch: the most recent time a fetch operation failed.

last_flush: the most recent time the flush function was called.

_timer: A Timer, handles calling the function that fetches the data.

Base.delete! — Method

Deletes all watcher data from storage backend. Also empties the buffer.

Base.deleteat! — Method

Delete watcher data from storage backend within the date range specified.

Base.get — Method

Returns the processed view of the watcher data. Accessible also as a view property of the watcher object.

Base.pop! — Method

Remove v from the things the watcher is fetching.

Base.push! — Method

Add v to the things the watcher is fetching.

Misc.start! — Method

Resets the watcher timer.

Misc.stop! — Method

Stops the watcher timer.

Watchers._fetch! — Method

Appends new data to the watcher buffer, returns true when new data is added, false otherwise.

Watchers._flush! — Method

May run after a successful fetch operation, according to the flush_interval. It spawns a task.

Watchers._get — Method

Returns the processed view of the watcher data.

Watchers._init! — Method

Function to run on watcher initialization, it runs before _load!.

Watchers._load! — Method

Called once on watcher creation, used to pre-fill the watcher buffer.

Watchers._pop! — Method

Same as _push! but for removing elements.

Watchers._process! — Method

Processes the watcher data, called everytime the watcher fetches new data.

Watchers._push! — Method

If the watcher manager a group of things that it is fetching, _push! should add an element to it.

Watchers._start! — Method

Executed before starting the timer.

Watchers._stop! — Method

Executed after the timer has been stopped.

Watchers._watcher — Function

Instantiate a watcher.

T: The type of the underlying CircularBuffer

len: length of the circular buffer.

start: If true(default), the watcher will start fetching asap.

asyncio vs threads

Both _fetch! and _flush! callbacks assume non-blocking asyncio like behaviour. If instead your functions require high computation, pass threads=true, you will have to ensure thread safety.

Watchers.allerror — Method

Get all logged watcher errors of type t.

Watchers.default_flusher — Method

Save function for watcher data, saves to the default DATA_PATH located lmdb instance using serialization.

Watchers.default_loader — Method

Load function for watcher data, loads from the default DATA_PATH located lmdb instance using serialization.

Watchers.default_process — Method

Processes the values of a watcher buffer into a dataframe. To be used with: default_init and default_get

Watchers.fetch! — Method

Fetches a new value from the watcher ignoring the timer. If reset is true the timer is reset and polling will resume after the watcher interval.

Watchers.flush! — Method

Flush the watcher. If wait is true, block until flush completes.

Watchers.isstale — Method

True if last available data entry is older than now() + fetch_interval + fetch_timeout.

Watchers.isstarted — Method

True if timer is running.

Watchers.isstopped — Method

True if timer is not running.

Watchers.lasterror — Method

Get the last logged watcher error of type t.

Watchers.lasterror — Method

Get the last logged watcher error.

Watchers.logerror — Function

Stores an error to the watcher log journal.

Watchers.pushnew! — Function

Helper function to push a new value to the watcher buffer if it is different from the last one.

Watchers.pushstart! — Method

Helper function to push a vector of values to the watcher buffer.

Watchers.WatchersImpls._fetchto! — Function

op: appendmax! or prependmax! If the watcher has attribute resync_noncontig set to true, preloaded data will be discarded if non contiguous.

Watchers.WatchersImpls._resolve — Function

This function checks that the date candidate is the correct next date to append to ohclv_dst.

Watchers.WatchersImpls._warmup! — Method

Checks if we can start processing data, after we are past the initial incomplete timeframe.

Watchers.WatchersImpls.ccxt_ohlcv_tickers_watcher — Method

OHLCV watcher based on exchange tickers data. This differs from the ohlcv watcher based on trades.

The OHLCV ticker watcher can monitor a group of symbols, while the trades watcher only one symbol per instance.

The OHLCV ticker watcher candles do not match 1:1 the exchange candles, since they rely on polled data.

The OHLCV ticker watcher is intended to be lazy. It won't pre-load/fetch data for all symbols, it will only

process new candles from the time it is started w.r.t. the timeframe provided.

The source price chooses which price to use to build the candles any of :last, :vwap, :bid, :ask (default :last).

To back-fill the view (DataFrame) of a particular symbol, call load!(watcher, symbol), which will fill the view up to the watcher view_capacity.

logfile: optional path to save errors.

Inaccurate volume

Since the volume data from the ticker is a daily rolling sum, the recorded volume is adjusted to be a fraction of it (using the timeframe as unit, see _meanvolume!). If this behaviour is not desired, it is better to use vwap as price source for building OHLC and ignore the volume column.

Watchers.WatchersImpls.ccxt_ohlcv_watcher — Method

Create a Watcher instance that tracks ohlcv for an exchange (ccxt).

On startup candles are initially loaded from storage (if any).

Then they are fastfowarded to the last available candle.

After which, fetching happen on trades

Once time crosses the timeframe, new candles are created from trades.

If The watcher is restarted, a new call for OHLCV data is made to re-fastfoward. If no trades happen during a timeframe, an empty candle for that timeframe is added. The view of the watcher SHOULD NOT have duplicate candles (same timestamp), and all the candles SHOULD be contiguous (the time difference between adjacent candles is always equal to the timeframe).

If these constraints are not met that's a bug.

Watcher data.

The data saved by the watcher on disk SHOULD NOT be relied upon to be contiguous, since the watcher doesn't ensure it, it only uses it to reduce the number of candles to fetch from the exchange at startup.

Watchers.WatchersImpls.ccxt_tickers_watcher — Method

Create a Watcher instance that tracks all markets for an exchange (ccxt).

Watchers.WatchersImpls.cg_derivatives_watcher — Method

Create a Watcher instance that tracks all the derivatives from an exchange.

Watchers.WatchersImpls.cg_ticker_watcher — Method

Create a Watcher instance that tracks the price of some currencies on an exchange (coingecko).

Watchers.WatchersImpls.cp_markets_watcher — Function

Create a Watcher instance that tracks all markets for an exchange (coinpaprika).

Watchers.WatchersImpls.cp_twitter_watcher — Function

Create a Watcher instance that tracks all markets for an exchange (coinpaprika).

Watchers.WatchersImpls.@append_dict_data — Macro

Defines a closure that appends new data on each symbol dataframe.