Watchers

Watchers.Watchers — Module

Watchers are data feeds, that keep track of stale data.

Watchers.Watcher — Type

Watchers manage data, they pull from somewhere, keep a cache in memory, and optionally flush periodically to persistent storage.

data: A CircularBuffer of default length 1000 of the watcher type parameter.

timer: A Timer, handles calling the function that fetches the data.

timeout: How much time to wait for the fetcher function.

interval: the Period with which the fetcher function will be called.

attempts: In cause of fetcher failure, tracks how many consecutive fails have occurred. It resets after a successful fetch operation.

last_try: the last time a fetch operation failed.

Watchers.Watcher3 — Type

Instantiate a watcher.

T: The type of the underlying CircularBuffer

len: length of the circular buffer.

fetcher: The input function that fetches data from somewhere, with signature () -> T.

flusher: Optional function that is called once every len successful updates, (AbstractVector) -> ()

asyncio vs threads

Both fetcher and flusher callbacks assume non-blocking asyncio like behaviour. If instead your functions require high computation, pass threads=true, you will have to ensure thread safety.

Warning

If using flushers, do not modify the input data (argument of the flusher callback), always make a copy.

Watchers.isstale — Method

True if last available data entry is older than now() + interval + timeout.